Refactor page composition into dedicated app component
