+++ app/page.tsx
@@ -1,2018 +1,5 @@
-"use client";
+import { JapaneseFoodApp } from "@/components/app/JapaneseFoodApp";
 
-import { useState, useRef, useLayoutEffect, useMemo, useEffect, type CSSProperties } from "react";
-import Image from "next/image";
-import { Button } from "@/components/ui/button";
-import { Card } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Badge } from "@/components/ui/badge";
-import {
-  ShoppingCart as ShoppingCartOutline,
-  Calendar as CalendarOutline,
-  Clock as ClockOutline,
-  User as UserOutline,
-  Search,
-  Heart } from
-"lucide-react";
-
-type Screen =
-"dashboard" |
-"catalog" |
-"catalogLanding" |
-"cart" |
-"order" |
-"history" |
-"profile" |
-"subscription";
-
-type SidebarNavKey = Screen | "subscription";
-
-type SidebarIconProps = {
-  fill: string;
-  stroke: string;
-};
-
-type SidebarIconComponent = (props: SidebarIconProps) => JSX.Element;
-
-interface Product {
-  id: number;
-  name: string;
-  description: string;
-  price: number;
-  image: string;
-  quantity: number;
-}
-
-type LandingCardContent = {
-  title: string;
-  renderIcon?: () => JSX.Element | null;
-};
-
-const sampleProducts: Product[] = [
-{
-  id: 1,
-  name: "トップリブ 特製",
-  description: "ナンバー 260g",
-  price: 350,
-  image: "/images/food-item.jpg",
-  quantity: 0
-},
-{
-  id: 2,
-  name: "ビーフステーキ",
-  description: "プレミアム 300g",
-  price: 450,
-  image: "/images/food-item.jpg",
-  quantity: 0
-},
-{
-  id: 3,
-  name: "チキンカツ",
-  description: "ジューシー 200g",
-  price: 280,
-  image: "/images/food-item.jpg",
-  quantity: 0
-},
-{
-  id: 4,
-  name: "ポークソテー",
-  description: "厚切り 250g",
-  price: 380,
-  image: "/images/food-item.jpg",
-  quantity: 0
-},
-{
-  id: 5,
-  name: "ハンバーグ",
-  description: "手作り 180g",
-  price: 320,
-  image: "/images/food-item.jpg",
-  quantity: 0
-},
-{
-  id: 6,
-  name: "サーモン焼き",
-  description: "ノルウェー産 220g",
-  price: 420,
-  image: "/images/food-item.jpg",
-  quantity: 0
-},
-{
-  id: 7,
-  name: "エビフライ",
-  description: "大ぶり 3本",
-  price: 390,
-  image: "/images/food-item.jpg",
-  quantity: 0
-},
-{
-  id: 8,
-  name: "唐揚げ",
-  description: "ジューシー 6個",
-  price: 300,
-  image: "/images/food-item.jpg",
-  quantity: 0
-}];
-
-const CATALOG_FILTER_BUTTONS = [
-  { label: "お気に入り", buttonDataOid: "jm:hia2", textDataOid: "btn-text-favorite" }
-];
-
-const LANDING_FILTER_BUTTONS = [
-  { label: "健康重視", buttonDataOid: ".zvc9j.", textDataOid: "btn-text-health" },
-  { label: "お気に入り", buttonDataOid: "jm:hia2", textDataOid: "btn-text-favorite" }
-];
-
-const FILTER_BUTTON_INACTIVE_STYLE: CSSProperties = {
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-  width: "max-content",
-  minWidth: "160px",
-  height: "60px",
-  padding: "0 24px",
-  flexShrink: 0,
-  borderRadius: "20px",
-  border: "2px solid #FDA900",
-  background: "var(--, #FFF)",
-  boxShadow: "4.5px 4.5px 0 0 #E4E2E2"
-};
-
-const FILTER_BUTTON_TEXT_STYLE: CSSProperties = {
-  color: "var(--, #101010)",
-  fontFamily: '"BIZ UDPGothic"',
-  fontSize: "32px",
-  fontStyle: "normal",
-  fontWeight: 700,
-  lineHeight: "normal",
-  letterSpacing: "1.664px"
-};
-
-const LandingIconDefault = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="100"
-    height="100"
-    viewBox="0 0 100 100"
-    fill="none">
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M23.1571 12.2211C27.6357 10.2783 31.1928 10.3569 34.0071 11.6354C36.9 12.9497 39.7714 15.8926 42.0714 21.1783C42.938 23.1926 43.5238 25.0497 43.8285 26.7497C44.2428 28.964 46.2357 30.5212 48.4857 30.3926C51.2428 30.2212 53.8357 30.9854 56.1857 32.9569C55.1142 33.6473 54.1238 34.4497 53.2142 35.364C50.9285 37.6497 49.3428 40.4354 48.3928 43.5926C47.9387 45.1081 47.6255 46.6623 47.4571 48.2354C47.2537 50.0879 47.2154 51.9548 47.3428 53.814C46.1333 53.814 44.9261 53.8212 43.7214 53.8354C42.8214 47.1997 41.2428 39.4354 34.7571 31.8426C34.3805 31.3833 33.9161 31.0037 33.3909 30.7262C32.8658 30.4486 32.2906 30.2786 31.699 30.2262C31.1073 30.1738 30.5112 30.2399 29.9454 30.4208C29.3797 30.6016 28.8557 30.8936 28.4042 31.2795C27.9527 31.6654 27.5827 32.1375 27.316 32.6682C27.0493 33.1989 26.8911 33.7775 26.8508 34.3701C26.8105 34.9627 26.8888 35.5573 27.0813 36.1193C27.2737 36.6812 27.5763 37.1991 27.9714 37.6426C32.5285 42.9783 33.8642 48.2283 34.7142 53.9854C28.5139 54.1252 22.3157 54.349 16.1214 54.6569C15.0499 49.4997 16.3642 45.7712 19.15 43.1212C19.9381 42.3712 20.4253 41.3593 20.52 40.2755C20.6147 39.1917 20.3104 38.1107 19.6642 37.2354C18.5191 35.6522 17.5581 33.9435 16.7999 32.1426C14.5 26.8569 14.3142 22.7569 15.3285 19.7426C16.3071 16.814 18.6857 14.1711 23.1571 12.2211ZM91.4285 56.6283C91.6904 55.0473 91.819 53.4021 91.8142 51.6926C91.8142 45.1712 89.9714 39.4497 85.8785 35.364C81.7857 31.2783 76.0642 29.4283 69.5428 29.4283C68.1142 29.4283 66.7261 29.5188 65.3785 29.6997C61.7714 25.1426 57.1357 22.2711 51.7357 21.5926C51.3218 20.2407 50.8281 18.9145 50.2571 17.6211C47.3999 11.0426 43.2357 6.02115 37.6999 3.49972C32.0857 0.956864 25.8357 1.32829 19.6071 4.03544C13.3714 6.73544 8.82852 11.0497 6.86424 16.8926C4.92852 22.6569 5.74995 29.1283 8.60709 35.6997C9.19281 37.0331 9.82614 38.2973 10.5071 39.4926C6.90709 44.464 6.04281 50.5783 7.52138 57.0926C5.64395 58.0185 4.09373 59.4952 3.07777 61.3254C2.06181 63.1557 1.62848 65.2524 1.83567 67.3354C2.79281 76.814 5.02138 85.0854 12.7071 90.6354C20.0714 95.9497 31.7857 98.214 50 98.214C68.2142 98.214 79.9214 95.9569 87.2928 90.6354C94.9785 85.0926 97.2071 76.8212 98.1642 67.3426C98.3838 65.0633 97.8419 62.7755 96.6232 60.837C95.4045 58.8984 93.5777 57.4184 91.4285 56.6283ZM56.2999 53.8426C65.1064 53.9236 73.9105 54.176 82.7071 54.5997C82.8261 53.7092 82.8857 52.7402 82.8857 51.6926C82.8857 46.814 81.5285 43.6426 79.5642 41.6783C77.5999 39.714 74.4214 38.3569 69.5428 38.3569C67.6714 38.3664 66.038 38.5521 64.6428 38.914L64.5428 38.9426C64.0219 39.0802 63.5092 39.2471 63.0071 39.4426C62.0547 39.8235 61.2214 40.2831 60.5071 40.8212C59.2694 41.7471 58.2824 42.9673 57.6357 44.3712C56.75 46.2069 56.2071 48.6069 56.2071 51.6926C56.2071 52.4497 56.238 53.1712 56.2999 53.8426Z"
-      fill="#101010"
-    />
-  </svg>
-);
-
-const LandingIconSecond = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="100"
-    height="100"
-    viewBox="0 0 100 100"
-    fill="none">
-    <path
-      d="M29.4729 52.3893H21.2503C21.6221 51.6143 22.2053 50.9602 22.9327 50.5024C23.6601 50.0446 24.5021 49.8017 25.3616 49.8017C26.2211 49.8017 27.0631 50.0446 27.7905 50.5024C28.518 50.9602 29.1012 51.6143 29.4729 52.3893ZM42.629 49.7932C42.6316 48.7271 42.7101 47.6625 42.864 46.6076C42.2256 46.7105 41.6164 46.948 41.0769 47.3045C40.5373 47.6609 40.0798 48.1281 39.7347 48.6749C39.3897 49.2218 39.1649 49.8358 39.0754 50.4762C38.9859 51.1166 39.0336 51.7687 39.2155 52.3893H42.7888C42.6877 51.5276 42.6374 50.6607 42.6384 49.7932H42.629ZM24.1657 32.991C21.1116 29.9369 17.0099 26.709 13.3919 26.709H13.0911L13.2815 30.468C15.0224 30.3834 17.8298 32.0891 20.9942 35.1664C21.1932 35.3591 21.3882 35.5526 21.5792 35.7469C22.0983 35.6348 22.627 35.5727 23.1579 35.5613C23.35 34.6527 23.6938 33.783 24.1751 32.9887L24.1657 32.991ZM28.9255 29.2932C29.431 28.6994 29.8837 28.0627 30.2786 27.3902C30.8765 26.3759 31.3431 25.2897 31.6673 24.1578C31.9902 23.0331 32.1678 21.8716 32.1958 20.7018C32.2206 19.6133 32.111 18.526 31.8692 17.4645C31.7625 17.0021 31.6315 16.5458 31.4767 16.0973C31.3483 15.7264 31.2026 15.3618 31.0399 15.0047C30.9264 14.7573 30.8033 14.5144 30.671 14.2766C30.6265 14.192 30.5794 14.1238 30.5536 14.0793L30.5112 14.0109L30.4409 14.0461C30.3938 14.0697 30.321 14.1051 30.2388 14.152C29.9982 14.2841 29.763 14.4259 29.5339 14.5771C29.2073 14.7938 28.8936 15.0291 28.5942 15.282C28.2276 15.5828 27.877 15.9027 27.544 16.2404C26.7824 17.0174 26.1196 17.8854 25.5706 18.8248C24.9816 19.8359 24.5245 20.9183 24.2104 22.0455C23.8874 23.1759 23.7089 24.3427 23.6794 25.518C23.6563 26.6063 23.766 27.6934 24.006 28.7553C24.1991 29.5976 24.4762 30.4185 24.8329 31.2057C24.9247 31.4123 25.014 31.5885 25.0938 31.7365C26.1225 30.5808 27.4524 29.734 28.9349 29.2908L28.9255 29.2932ZM95.803 20.7184C95.7585 20.6924 95.6878 20.65 95.6034 20.6055C95.3634 20.4761 95.1172 20.3585 94.8657 20.2531C94.5532 20.1215 94.1821 19.9758 93.764 19.842C93.3117 19.698 92.8521 19.5781 92.3872 19.4826C91.32 19.2671 90.2303 19.1843 89.1429 19.2359C87.9735 19.2916 86.816 19.4969 85.6987 19.8467C84.5802 20.1931 83.5099 20.6793 82.5132 21.2939C81.587 21.8671 80.7342 22.5511 79.9735 23.3307L79.8677 23.4436C80.9489 24.7226 81.7844 26.1905 82.3321 27.7732C83.0647 27.8613 83.8034 27.8888 84.5405 27.8555C85.7102 27.8001 86.8682 27.5973 87.987 27.2518C89.1067 26.9105 90.178 26.4273 91.1749 25.8139C92.1021 25.2422 92.9551 24.5581 93.7146 23.7771C94.0465 23.4361 94.3602 23.0777 94.6542 22.7035C94.9245 22.3559 95.1522 22.0316 95.3333 21.7402C95.5142 21.4488 95.6526 21.2094 95.7421 21.0354C95.789 20.9508 95.822 20.8758 95.8431 20.8287L95.8782 20.7559L95.803 20.7184ZM79.0761 7.81602C78.9915 7.48457 78.9093 7.17227 78.8835 7.03125L75.1763 7.7125C75.2255 7.98281 75.3101 8.30684 75.4253 8.75332C75.9749 10.8818 77.2413 15.7871 75.7144 20.1264C76.8692 20.7271 77.9324 21.4897 78.8716 22.391C81.369 16.741 79.7526 10.4707 79.0667 7.81582L79.0761 7.81602ZM35.1911 92.9688H59.1399C63.221 92.9688 67.2621 92.165 71.0325 90.6033C74.8029 89.0416 78.2288 86.7525 81.1145 83.8668C84.0002 80.981 86.2893 77.5551 87.851 73.7847C89.4128 70.0143 90.2166 65.9732 90.2165 61.8922V56.148H4.12158V61.8945C4.1221 70.135 7.3955 78.0378 13.2219 83.8651C19.0483 89.6924 26.9507 92.967 35.1911 92.9688ZM83.0745 49.7932C83.0793 45.0733 81.2574 40.5347 77.9905 37.1281C78.7451 35.8209 79.1889 34.3577 79.2878 32.8516C79.3866 31.3455 79.1378 29.8369 78.5605 28.4423C77.9832 27.0477 77.093 25.8045 75.9587 24.8088C74.8243 23.8132 73.4762 23.0917 72.0185 22.7002C70.5608 22.3087 69.0326 22.2577 67.552 22.551C66.0715 22.8443 64.6782 23.4742 63.4799 24.3919C62.2816 25.3096 61.3104 26.4906 60.6413 27.8436C59.9723 29.1965 59.6233 30.6852 59.6214 32.1945C55.3778 33.4247 51.7215 36.1468 49.3261 39.8594C46.9306 43.572 45.9573 48.0252 46.5853 52.3986H82.8819C83.0073 51.5357 83.0716 50.6651 83.0745 49.7932ZM35.278 51.1111C35.2771 49.0919 36.014 47.1419 37.3501 45.6279C36.9128 44.7977 36.2303 44.1224 35.3956 43.6938C34.5608 43.2652 33.6143 43.1041 32.6848 43.2325C31.7553 43.3608 30.8879 43.7724 30.2005 44.4111C29.5132 45.0499 29.0393 45.8849 28.8433 46.8025C29.9703 47.3246 30.9669 48.0912 31.7605 49.0466C32.5542 50.0021 33.125 51.1223 33.4315 52.326C34.0845 52.306 34.7254 52.1456 35.3108 51.8559C35.2891 51.6082 35.2782 51.3597 35.278 51.1111ZM19.7726 48.1887C19.2013 47.1393 18.2416 46.3556 17.0992 46.0056C15.9569 45.6556 14.7228 45.7672 13.6618 46.3164C12.6008 46.8657 11.7972 47.8089 11.4235 48.9437C11.0498 50.0786 11.1357 51.3147 11.6628 52.3869H17.2659C17.6623 50.7686 18.5346 49.3063 19.7702 48.1887H19.7726ZM35.3673 35.1641C34.936 34.3334 34.2597 33.6554 33.4302 33.2219C32.6006 32.7885 31.6579 32.6205 30.7297 32.7408C29.8014 32.861 28.9326 33.2637 28.241 33.8943C27.5493 34.5249 27.0682 35.3529 26.8628 36.266C28.4852 37.0273 29.8207 38.2887 30.6733 39.865C31.3952 39.6226 32.1471 39.4811 32.9077 39.4445C33.2755 37.8023 34.1321 36.3099 35.3647 35.1641H35.3673ZM41.0313 36.6912C40.0174 36.6892 39.0315 37.0247 38.2293 37.6449C37.4271 38.265 36.8541 39.1346 36.6007 40.1164C38.182 40.7998 39.5142 41.9541 40.4159 43.4219C41.4265 42.9995 42.511 42.7824 43.6063 42.783H43.7731C44.1583 41.6263 44.64 40.5041 45.2132 39.4281C44.8576 38.6149 44.2727 37.9229 43.53 37.4368C42.7874 36.9508 41.9191 36.6916 41.0315 36.6912H41.0313ZM27.4524 41.8479C27.0429 41.0176 26.3904 40.3315 25.5817 39.8808C24.773 39.4301 23.8463 39.2361 22.9247 39.3246C22.0031 39.4131 21.1303 39.7798 20.4221 40.3761C19.7139 40.9724 19.2038 41.77 18.9597 42.6631C20.7033 43.3987 22.1466 44.7028 23.0546 46.3633C23.7426 46.1655 24.4533 46.0573 25.169 46.0414C25.5041 44.4506 26.298 42.9925 27.4524 41.8479Z"
-      fill="#101010"
-    />
-  </svg>
-);
-
-const LandingIconThird = () => (
-  <svg 
-    xmlns="http://www.w3.org/2000/svg" 
-    width="82" 
-    height="92" 
-    viewBox="0 0 82 92" 
-    fill="none">
-    <path 
-      fill-rule="evenodd" 
-      clip-rule="evenodd" 
-      d="M57.5082 0.640962C61.8415 0.849295 65.6665 1.7743 68.6457 2.87013C71.4582 3.90763 73.979 5.26596 75.354 6.64513C76.729 8.02013 78.0915 10.541 79.1248 13.3535C80.4305 16.931 81.1836 20.6866 81.3582 24.491C81.7123 31.9493 79.9207 40.8535 73.3582 49.3618L73.2832 49.5576L67.2957 61.541C66.9208 62.2908 66.3286 62.9101 65.5962 63.318C64.8639 63.7259 64.0255 63.9034 63.1907 63.8273C62.3559 63.7513 61.5634 63.4252 60.9169 62.8917C60.2703 62.3581 59.7997 61.642 59.5665 60.8368C54.1695 63.5256 48.3206 65.1899 42.3165 65.7451L40.1915 65.9201L37.7623 66.0826L33.6957 66.3035L29.3123 66.5118C27.2165 66.6118 25.0873 66.7118 23.029 66.8201C22.8645 69.1252 22.6087 71.4228 22.2623 73.7076C22.0721 75.0582 21.7497 76.3868 21.2998 77.6743C20.504 79.766 18.8123 81.5201 16.429 81.5201C13.2957 81.5201 10.4457 79.3118 8.12484 77.116L7.00817 76.0326L5.42484 74.4451C3.07484 72.0285 0.479004 68.966 0.479004 65.5701C0.479004 63.6535 1.5915 62.4118 2.29567 61.8326C3.53317 60.8118 5.13317 60.3326 6.72067 60.0118L8.28734 59.7368C10.5735 59.3903 12.8726 59.1345 15.179 58.9701L15.7623 46.9035L15.9165 44.2368L16.0748 41.8076C16.1304 41.0437 16.1901 40.3354 16.254 39.6826C16.829 33.6243 18.5082 27.6826 21.1623 22.4326C20.3574 22.199 19.6416 21.7279 19.1085 21.0811C18.5754 20.4343 18.2498 19.6417 18.1742 18.807C18.0986 17.9722 18.2765 17.134 18.6847 16.402C19.093 15.6699 19.7125 15.0779 20.4623 14.7035L32.2457 8.80763C32.3733 8.74544 32.504 8.68981 32.6373 8.64096C41.1415 2.07846 50.0498 0.282628 57.5082 0.640962ZM60.6248 21.3743C59.4529 20.2018 57.8632 19.5429 56.2055 19.5425C54.5477 19.5421 52.9577 20.2003 51.7852 21.3722C50.6128 22.5441 49.9539 24.1338 49.9535 25.7916C49.9531 27.4493 50.6112 29.0393 51.7832 30.2118C52.9614 31.3508 54.5399 31.9816 56.1786 31.9681C57.8173 31.9546 59.3852 31.2981 60.5446 30.1398C61.7039 28.9816 62.362 27.4143 62.377 25.7756C62.392 24.1368 61.7628 22.5536 60.6248 21.3743Z" 
-      fill="#101010"/>
-  </svg>
-);
-
-const LandingIconFourth = () => (
-  <svg 
-    xmlns="http://www.w3.org/2000/svg" 
-    width="84" 
-    height="84" 
-    viewBox="0 0 84 84" 
-    fill="none">
-    <path 
-      d="M75.9999 45.0419C78.7709 42.2748 80.8843 38.9201 82.1839 35.226C83.4835 31.5319 83.936 27.5929 83.5079 23.7003C83.0798 19.8077 81.7822 16.0612 79.7109 12.7378C77.6396 9.41434 74.8476 6.5991 71.5416 4.50024C63.1666 -1.04143 51.9582 -1.0831 43.4999 4.37524C36.1666 9.08357 32.1249 16.4169 31.4166 24.0002C30.8749 29.5002 28.7916 34.6252 24.9582 38.4586L24.8332 38.5836C19.9999 43.4169 19.9999 50.7919 24.5416 55.2919L28.6666 59.4169C30.8496 61.5985 33.8095 62.8239 36.8957 62.8239C39.9819 62.8239 42.9419 61.5985 45.1249 59.4169C49.1666 55.3752 54.4999 53.1669 60.2915 52.5419C65.9999 51.9169 71.5832 49.4169 75.9999 45.0419ZM18.0832 74.7502C19.2082 77.0836 18.8332 79.9169 16.8749 81.8336C16.0267 82.6945 14.9457 83.2892 13.7645 83.5447C12.5833 83.8001 11.3532 83.7053 10.2251 83.2717C9.09695 82.8382 8.11991 82.0848 7.41378 81.104C6.70766 80.1232 6.30316 78.9576 6.24989 77.7502C5.04252 77.697 3.87694 77.2925 2.89614 76.5863C1.91534 75.8802 1.16196 74.9032 0.728414 73.7751C0.294868 72.647 0.200009 71.4169 0.455472 70.2356C0.710935 69.0544 1.30561 67.9734 2.16656 67.1252C4.08322 65.2086 6.95822 64.7919 9.24989 65.9169L19.5832 55.7919C20.1666 56.5836 20.8332 57.5002 21.5832 58.2502L25.7082 62.3752C26.5832 63.2086 27.4166 63.9169 28.4999 64.5419L18.0832 74.7502Z" 
-      fill="#101010"/>
-  </svg>
-);
-
-const LandingIconFifth = () => (
-  <svg 
-    xmlns="http://www.w3.org/2000/svg" 
-    width="84" 
-    height="60" 
-    viewBox="0 0 84 60" 
-    fill="none">
-    <path 
-      d="M58.6668 25.833V0.833008H75.3335C77.6252 0.833008 79.5877 1.64967 81.221 3.28301C82.8543 4.91634 83.6696 6.87745 83.6668 9.16634V25.833H58.6668ZM8.66683 59.1663C6.37516 59.1663 4.41405 58.3511 2.7835 56.7205C1.15294 55.09 0.336274 53.1274 0.333496 50.833V9.16634C0.333496 6.87467 1.15016 4.91356 2.7835 3.28301C4.41683 1.65245 6.37794 0.835786 8.66683 0.833008H50.3335V59.1663H8.66683ZM25.3335 36.2497C27.0696 36.2497 28.546 35.6427 29.7627 34.4288C30.9793 33.215 31.5863 31.7386 31.5835 29.9997C31.5807 28.2608 30.9738 26.7858 29.7627 25.5747C28.5516 24.3636 27.0752 23.7552 25.3335 23.7497C23.5918 23.7441 22.1168 24.3524 20.9085 25.5747C19.7002 26.7969 19.0918 28.2719 19.0835 29.9997C19.0752 31.7275 19.6835 33.2038 20.9085 34.4288C22.1335 35.6538 23.6085 36.2608 25.3335 36.2497ZM58.6668 59.1663V34.1663H83.6668V50.833C83.6668 53.1247 82.8516 55.0872 81.221 56.7205C79.5904 58.3538 77.6279 59.1691 75.3335 59.1663H58.6668Z" 
-      fill="#101010"/>
-  </svg>
-);
-
-const LandingIconSixth = () => (
-  <svg 
-    xmlns="http://www.w3.org/2000/svg" 
-    width="76" 
-    height="76" 
-    viewBox="0 0 76 76" 
-    fill="none">
-    <path 
-      d="M67.1667 10.0833C72.1667 12.1667 75.5 17.5833 75.5 23.4167C75.5 52.1667 52.1667 75.5 23.4167 75.5C17.5833 75.5 12.5833 72.1667 10.0833 67.1667L0.5 73.4167V48.4167L10.0833 54.6667C12.5833 49.6667 17.5833 46.3333 23.4167 46.3333C35.9167 46.3333 46.3333 35.9167 46.3333 23.4167C46.3333 17.5833 49.6667 12.5833 54.6667 10.0833L48.4167 0.5H73.4167L67.1667 10.0833Z" 
-      fill="#101010"/>
-  </svg>
-);
-
-const LandingIconSeventh = () => (
-  <div className="flex h-[100px] w-[100px] items-center justify-center">
-    <Image
-      src="/images/DairyAndEggs.png"
-      alt="Dairy and Eggs"
-      width={100}
-      height={100}
-    />
-  </div>
-);
-
-/**
- * Update genre card titles/icons by editing LANDING_CARD_CONTENT_PAGES.
- * Use null for icon to keep the dashed placeholder.
- */
-const LANDING_CARD_CONTENT_PAGES: LandingCardContent[][] = [
-[
-  { title: "ジャンル 1", renderIcon: () => <LandingIconDefault /> },
-  { title: "ジャンル 2", renderIcon: () => <LandingIconSecond /> },
-  { title: "ジャンル 3", renderIcon: () => <LandingIconThird /> },
-  { title: "ジャンル 4", renderIcon: () => <LandingIconFourth /> },
-  { title: "ジャンル 5", renderIcon: () => <LandingIconFifth /> },
-  { title: "ジャンル 6", renderIcon: () => <LandingIconSixth /> },
-  { title: "ジャンル 7", renderIcon: () => <LandingIconSeventh /> },
-  { title: "ジャンル 8" }
-],
-[
-  { title: "ジャンル 9" },
-  { title: "ジャンル 10" },
-  { title: "ジャンル 11" },
-  { title: "ジャンル 12" }
-]
-];
-
-const SIDEBAR_ACTIVE_BG =
-  "url(\"data:image/svg+xml,%3Csvg%20xmlns=%22http://www.w3.org/2000/svg%22%20width=%22192%22%20height=%22100%22%20viewBox=%220%200%20192%20100%22%20fill=%22none%22%3E%3Cpath%20d=%22M0%2040.0005C0%2028.9548%208.9543%2020.0005%2020%2020.0005H192V79.2693H20C8.95431%2079.2693%200%2070.315%200%2059.2693V40.0005Z%22%20fill=%22white%22/%3E%3Cpath%20d=%22M192%2079.2695H172C183.046%2079.2697%20192%2088.224%20192%2099.2695V79.2695Z%22%20fill=%22white%22/%3E%3Cpath%20d=%22M192%2020H172C183.046%2019.9998%20192%2011.0456%20192%200V20Z%22%20fill=%22white%22/%3E%3C/svg%3E\")";
-
-const SIDEBAR_LABEL_BASE_STYLE = {
-  display: "flex",
-  width: "118px",
-  height: "52.616px",
-  justifyContent: "center",
-  alignItems: "center",
-  gap: "4.5px",
-  color: "#ffffff",
-  textAlign: "center" as const,
-  fontFamily: '"BIZ UDPGothic"',
-  fontSize: "20px",
-  fontStyle: "normal" as const,
-  fontWeight: 700,
-  lineHeight: "100%",
-  letterSpacing: "1.04px"
-};
-
-const SIDEBAR_LABEL_SELECTED_STYLE = {
-  color: "#101010",
-  textAlign: "center" as const,
-  fontFamily: '"BIZ UDPGothic"',
-  fontSize: "20px",
-  fontStyle: "normal" as const,
-  fontWeight: 700,
-  lineHeight: "100%",
-  letterSpacing: "1.04px"
-};
-
-const SIDEBAR_ICON_STYLE = {
-  display: "flex",
-  width: "31.5px",
-  height: "34.616px",
-  flexShrink: 0,
-  justifyContent: "center",
-  alignItems: "center"
-};
-
-const SidebarHomeIcon: SidebarIconComponent = ({ fill, stroke }) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 34 37"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M16.7432 0.5C17.5497 0.5 18.2899 0.711917 18.9521 1.13379L19.2314 1.32715L19.2324 1.32812L31.3486 10.501H31.3477C31.8648 10.8813 32.2695 11.3663 32.5596 11.9492C32.8525 12.5333 33 13.1558 33 13.8096V33.3662C33 34.1093 32.7175 34.76 32.1807 35.2969C31.6438 35.8337 30.9931 36.1162 30.25 36.1162H21.5107C20.8695 36.1162 20.308 35.8948 19.8613 35.4482L19.8604 35.4473C19.4147 34.9997 19.1934 34.4382 19.1934 33.7979V22.877H14.3066V33.7979C14.3066 34.4393 14.0854 35.0015 13.6387 35.4482C13.1923 35.8945 12.6315 36.1162 11.9912 36.1162H3.25C2.50693 36.1162 1.85619 35.8337 1.31934 35.2969C0.78249 34.76 0.5 34.1093 0.5 33.3662V13.8096C0.5 13.1558 0.647519 12.5333 0.94043 11.9492C1.23298 11.3659 1.63907 10.8799 2.15723 10.499L14.2676 1.32812C14.9908 0.776359 15.822 0.500017 16.7432 0.5ZM3.75 32.8662H11.0566V21.9434C11.0566 21.3023 11.2803 20.7412 11.7285 20.2949C12.1745 19.8478 12.7347 19.625 13.375 19.625H20.125C20.7661 19.625 21.3269 19.8484 21.7744 20.2959C22.2208 20.7423 22.4434 21.303 22.4434 21.9434V32.8662H29.75V13.8096C29.75 13.6539 29.7172 13.5251 29.6582 13.4141C29.5923 13.2901 29.5017 13.179 29.3818 13.0801L17.2725 3.95312L17.2588 3.94238L17.2451 3.93066C17.1089 3.81226 16.9523 3.75195 16.75 3.75195C16.5476 3.75195 16.392 3.81216 16.2578 3.92969L16.2441 3.94238L16.2295 3.95312L4.12891 13.0723C4.00367 13.1777 3.90828 13.2933 3.83887 13.4199C3.78206 13.5236 3.74904 13.6482 3.75 13.8057V32.8662Z"
-      fill={fill}
-      stroke={stroke}
-    />
-  </svg>
-);
-
-const SidebarCatalogIcon: SidebarIconComponent = ({ fill }) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 40 41"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M16.914 1.97034C17.2738 1.56623 17.7201 1.23721 18.2241 1.00451C18.728 0.7718 19.2785 0.640574 19.8399 0.619293C20.4014 0.598011 20.9614 0.687145 21.4838 0.880947C22.0062 1.07475 22.4794 1.36892 22.8729 1.74447L23.09 1.97034L30.997 10.8507H36.9981C37.42 10.8498 37.8373 10.9325 38.2228 11.0931C38.6082 11.2538 38.9531 11.4888 39.2347 11.7829C39.5164 12.077 39.7285 12.4235 39.8572 12.7996C39.9859 13.1758 40.0283 13.5732 39.9816 13.9658L39.7966 15.4076L39.5956 16.8004L39.4307 17.8432L39.2397 18.9631L39.0226 20.1395L38.7773 21.3592L38.5039 22.6033C38.3592 23.2307 38.2037 23.8575 38.0375 24.4837C37.5925 26.1363 37.0338 27.7603 36.3648 29.3454L35.9205 30.3618L35.4903 31.2804L35.0842 32.0973L34.8952 32.4624L34.3946 33.3847C33.7513 34.5329 32.537 35.2105 31.2523 35.3083L30.9307 35.3196H9.05531C8.35247 35.3246 7.6614 35.1507 7.05591 34.8165C6.45042 34.4823 5.95324 34.0005 5.61749 33.4223L5.15108 32.5753L4.7892 31.8789L4.59821 31.4931L4.19613 30.6479C3.27797 28.6519 2.53304 26.5898 1.96858 24.4818C1.85944 24.0725 1.75489 23.6621 1.65496 23.2508L1.36948 22.0311L1.11416 20.8378L0.888989 19.684L0.687947 18.5791L0.513041 17.5364L0.36427 16.5708L0.183332 15.2852L0.0506448 14.2387L0.0164677 13.9507C-0.0251189 13.5809 0.0122512 13.2071 0.126404 12.851C0.240557 12.4948 0.429213 12.1635 0.681397 11.8762C0.933582 11.589 1.24426 11.3516 1.59536 11.1778C1.94646 11.004 2.33096 10.8973 2.72651 10.8639L2.99993 10.8507H9.00907L16.914 1.97034ZM35.8502 14.6151H4.14587L4.27051 15.5261L4.4233 16.5576L4.60826 17.6945L4.82539 18.9141C4.90581 19.332 4.99024 19.7599 5.0787 20.1978L5.36619 21.5286C5.51898 22.2024 5.68786 22.8875 5.86879 23.5689C6.379 25.4715 7.05129 27.3327 7.87921 29.1346L8.27326 29.9647L8.63915 30.6875L8.96685 31.2954L9.1116 31.5552H30.8784L31.188 30.9849L31.5438 30.2979L31.9339 29.5017C32.6717 27.9488 33.5161 25.8916 34.1353 23.5689C34.3866 22.6278 34.6097 21.6716 34.8088 20.7324L35.0902 19.3433L35.3315 18.0088L35.438 17.3707L35.629 16.1717L35.7838 15.0989L35.8502 14.6151ZM15.9531 19.9531L16.9583 25.5998C17.0285 26.0848 16.8941 26.5762 16.5838 26.97C16.2735 27.3638 15.8115 27.629 15.2959 27.7093C14.7803 27.7896 14.2515 27.6787 13.8218 27.4001C13.392 27.1216 13.095 26.6973 12.9937 26.2172L11.9885 20.5705C11.9364 20.3233 11.9378 20.0689 11.9926 19.8222C12.0475 19.5755 12.1547 19.3417 12.3079 19.1345C12.461 18.9274 12.6571 18.7511 12.8843 18.6162C13.1116 18.4813 13.3655 18.3905 13.631 18.3492C13.8964 18.3078 14.1681 18.3168 14.4298 18.3756C14.6915 18.4343 14.9379 18.5417 15.1544 18.6913C15.371 18.8408 15.5532 19.0296 15.6904 19.2464C15.8276 19.4631 15.9169 19.7035 15.9531 19.9531ZM26.1298 18.3796L26.363 18.4041C26.8496 18.4792 27.2897 18.7193 27.5996 19.0785C27.9095 19.4377 28.0676 19.891 28.0437 20.3522L28.0176 20.5705L27.0124 26.2172C26.9271 26.6898 26.6524 27.1142 26.2444 27.4038C25.8363 27.6933 25.3256 27.8262 24.8165 27.7753C24.3074 27.7244 23.8382 27.4935 23.5047 27.1298C23.1712 26.7661 22.9984 26.297 23.0217 25.8182L23.0478 25.5998L24.053 19.9531C24.1333 19.4982 24.389 19.0867 24.7719 18.7967C25.1547 18.5066 25.6379 18.3582 26.1298 18.3796ZM20.002 4.37959L14.2422 10.8507H25.7619L20.002 4.37959Z"
-      fill={fill}
-    />
-  </svg>
-);
-
-const SidebarCartIcon: SidebarIconComponent = ({ fill }) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 38 45"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M11.9445 13.7402C11.7807 13.5242 11.685 13.2643 11.6696 12.9936C11.6542 12.723 11.7198 12.4539 11.858 12.2207C11.9961 11.9874 12.2007 11.8007 12.4454 11.6842C12.6902 11.5677 12.9641 11.5267 13.2323 11.5665C13.5909 11.618 13.9144 11.8098 14.1318 12.0997L19.8735 19.7554V1.98329C19.8735 1.62072 20.0176 1.273 20.274 1.01662C20.5303 0.760243 20.8781 0.616211 21.2406 0.616211C21.6032 0.616211 21.9509 0.760243 22.2073 1.01662C22.4637 1.273 22.6077 1.62072 22.6077 1.98329V19.7554L28.3495 12.0997C28.5123 11.8846 28.7351 11.7224 28.9897 11.6335C29.2444 11.5445 29.5197 11.5327 29.7811 11.5996C30.0424 11.6664 30.2782 11.809 30.4589 12.0094C30.6395 12.2097 30.757 12.459 30.7965 12.7258C30.822 12.9037 30.8122 13.0848 30.7676 13.2589C30.7231 13.4329 30.6446 13.5965 30.5368 13.7402L22.3343 24.6769C22.3179 24.6988 22.2878 24.7097 22.2687 24.7316C22.1776 24.8384 22.0706 24.9307 21.9515 25.005C21.9092 25.0405 21.8634 25.0716 21.8148 25.0979C21.6365 25.187 21.4399 25.2333 21.2406 25.2333C21.0413 25.2333 20.8448 25.187 20.6665 25.0979C20.6179 25.0716 20.5721 25.0405 20.5297 25.005C20.4106 24.9307 20.3036 24.8384 20.2126 24.7316C20.1934 24.7097 20.1634 24.6988 20.147 24.6769L11.9445 13.7402ZM13.7217 40.9452C13.7229 41.6213 13.5229 42.2824 13.1469 42.8443C12.771 43.4063 12.2363 43.8435 11.6109 44.1004C11.1442 44.2926 10.6407 44.3792 10.1365 44.354C9.63241 44.3288 9.14008 44.1925 8.69483 43.9547C8.24958 43.7169 7.86243 43.3836 7.5611 42.9786C7.25977 42.5737 7.05173 42.1071 6.95187 41.6123C6.85675 41.1438 6.86097 40.6605 6.96427 40.1938C7.06757 39.727 7.26765 39.2871 7.55158 38.9025C7.83551 38.5178 8.19697 38.1971 8.61261 37.9609C9.02825 37.7247 9.48882 37.5783 9.96457 37.5312C10.4403 37.4841 10.9206 37.5373 11.3745 37.6875C11.8284 37.8376 12.2457 38.0813 12.5996 38.4028C12.9534 38.7243 13.2358 39.1164 13.4287 39.5539C13.6215 39.9913 13.7204 40.4644 13.7189 40.9424L13.7217 40.9452ZM35.595 40.9452C35.5963 41.6213 35.3962 42.2824 35.0203 42.8443C34.6444 43.4063 34.1096 43.8435 33.4842 44.1004C33.0175 44.2926 32.514 44.3792 32.0099 44.354C31.5057 44.3288 31.0134 44.1925 30.5682 43.9547C30.1229 43.7169 29.7358 43.3836 29.4344 42.9786C29.1331 42.5737 28.9251 42.1071 28.8252 41.6123C28.7344 41.1457 28.7421 40.6653 28.8476 40.2019C28.9532 39.7384 29.1543 39.3021 29.4382 38.9208C29.7221 38.5395 30.0824 38.2217 30.4961 37.9877C30.9099 37.7537 31.3679 37.6087 31.8409 37.5618C32.314 37.515 32.7915 37.5675 33.2431 37.7158C33.6947 37.8642 34.1104 38.1051 34.4635 38.4233C34.8166 38.7416 35.0994 39.13 35.2937 39.5638C35.4881 39.9975 35.5898 40.4671 35.5923 40.9424L35.595 40.9452ZM36.2785 15.6541C35.916 15.6541 35.5682 15.7982 35.3119 16.0545C35.0555 16.3109 34.9115 16.6586 34.9115 17.0212L33.5444 27.9579H9.18295L6.17537 9.96706C6.12253 9.64923 5.95914 9.36027 5.71402 9.15117C5.46891 8.94207 5.15781 8.82626 4.83562 8.82418H2.10146C1.73889 8.82418 1.39116 8.96821 1.13478 9.22459C0.878407 9.48097 0.734375 9.82869 0.734375 10.1913C0.734375 10.5538 0.878407 10.9016 1.13478 11.1579C1.39116 11.4143 1.73889 11.5583 2.10146 11.5583H3.67634L7.5862 35.0175C7.63717 35.3377 7.80116 35.629 8.04845 35.8386C8.29573 36.0483 8.60994 36.1624 8.93414 36.1604H33.5416C33.9042 36.1604 34.2519 36.0163 34.5083 35.76C34.7647 35.5036 34.9087 35.1559 34.9087 34.7933C34.9087 34.4307 34.7647 34.083 34.5083 33.8266C34.2519 33.5702 33.9042 33.4262 33.5416 33.4262H10.0825L9.62588 30.692H33.5225C34.1888 30.6919 34.8321 30.4484 35.3315 30.0073C35.831 29.5663 36.1522 28.9581 36.2348 28.2969L37.6237 17.0321C37.6237 16.6696 37.4797 16.3219 37.2233 16.0655C36.967 15.8091 36.6192 15.6651 36.2567 15.6651L36.2785 15.6541Z"
-      fill={fill}
-    />
-  </svg>
-);
-
-const SidebarHistoryIcon: SidebarIconComponent = ({ fill }) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 36 35"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M17.9455 34.5179C13.93 34.5179 10.3766 33.2993 7.28537 30.8619C4.1941 28.4246 2.18636 25.3091 1.26216 21.5155C1.13469 21.0374 1.23029 20.5996 1.54898 20.2018C1.86767 19.8041 2.2979 19.5727 2.83967 19.5077C3.34957 19.444 3.81167 19.5396 4.22596 19.7945C4.64026 20.0495 4.92708 20.4319 5.08642 20.9418C5.85127 23.81 7.42878 26.1524 9.81894 27.9689C12.2091 29.7854 14.918 30.6937 17.9455 30.6937C21.6741 30.6937 24.8374 29.3953 27.4354 26.7987C30.0333 24.202 31.3317 21.0387 31.3304 17.3088C31.3291 13.5788 30.0308 10.4162 27.4354 7.82079C24.84 5.22539 21.6767 3.92642 17.9455 3.92387C15.7465 3.92387 13.691 4.43377 11.7789 5.45357C9.86675 6.47338 8.25737 7.8756 6.95075 9.66026H10.297C10.8387 9.66026 11.2932 9.84382 11.6603 10.211C12.0275 10.5781 12.2104 11.0319 12.2091 11.5724C12.2078 12.1129 12.0243 12.5673 11.6584 12.9357C11.2926 13.3041 10.8387 13.4871 10.297 13.4845H2.64846C2.10669 13.4845 1.65288 13.301 1.28702 12.9338C0.921168 12.5667 0.737603 12.1129 0.736328 11.5724V3.92387C0.736328 3.3821 0.919893 2.92829 1.28702 2.56243C1.65415 2.19658 2.10796 2.01301 2.64846 2.01174C3.18895 2.01046 3.6434 2.19403 4.01181 2.56243C4.38021 2.93084 4.56314 3.38465 4.56059 3.92387V6.50524C6.1859 4.46564 8.17005 2.88813 10.513 1.77272C12.856 0.657314 15.3335 0.0996094 17.9455 0.0996094C20.3357 0.0996094 22.5748 0.554059 24.6628 1.46296C26.7509 2.37186 28.5674 3.59817 30.1124 5.14189C31.6574 6.68562 32.8843 8.50215 33.7932 10.5915C34.7021 12.6808 35.1559 14.9199 35.1547 17.3088C35.1534 19.6977 34.6996 21.9368 33.7932 24.0261C32.8869 26.1154 31.6599 27.9319 30.1124 29.4757C28.5648 31.0194 26.7483 32.2463 24.6628 33.1565C22.5773 34.0667 20.3382 34.5205 17.9455 34.5179ZM19.8576 16.5439L24.6379 21.3242C24.9885 21.6748 25.1638 22.121 25.1638 22.6627C25.1638 23.2045 24.9885 23.6507 24.6379 24.0012C24.2874 24.3518 23.8412 24.5271 23.2995 24.5271C22.7577 24.5271 22.3115 24.3518 21.961 24.0012L16.607 18.6473C16.4158 18.4561 16.2724 18.2413 16.1768 18.0029C16.0812 17.7645 16.0334 17.5172 16.0334 17.261V9.66026C16.0334 9.11849 16.2169 8.66468 16.5841 8.29882C16.9512 7.93297 17.405 7.7494 17.9455 7.74813C18.486 7.74685 18.9404 7.93042 19.3088 8.29882C19.6772 8.66722 19.8602 9.12104 19.8576 9.66026V16.5439Z"
-      fill={fill}
-    />
-  </svg>
-);
-
-const SidebarSubscriptionIcon: SidebarIconComponent = ({ fill }) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 40 41"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M34 4.25391H30V2.25391C30 1.72347 29.7893 1.21477 29.4142 0.839693C29.0391 0.46462 28.5304 0.253906 28 0.253906C27.4696 0.253906 26.9609 0.46462 26.5858 0.839693C26.2107 1.21477 26 1.72347 26 2.25391V4.25391H14V2.25391C14 1.72347 13.7893 1.21477 13.4142 0.839693C13.0391 0.46462 12.5304 0.253906 12 0.253906C11.4696 0.253906 10.9609 0.46462 10.5858 0.839693C10.2107 1.21477 10 1.72347 10 2.25391V4.25391H6C4.4087 4.25391 2.88258 4.88605 1.75736 6.01127C0.632141 7.13648 0 8.66261 0 10.2539V34.2539C0 35.8452 0.632141 37.3713 1.75736 38.4965C2.88258 39.6218 4.4087 40.2539 6 40.2539H34C35.5913 40.2539 37.1174 39.6218 38.2426 38.4965C39.3679 37.3713 40 35.8452 40 34.2539V10.2539C40 8.66261 39.3679 7.13648 38.2426 6.01127C37.1174 4.88605 35.5913 4.25391 34 4.25391ZM36 34.2539C36 34.7843 35.7893 35.293 35.4142 35.6681C35.0391 36.0432 34.5304 36.2539 34 36.2539H6C5.46957 36.2539 4.96086 36.0432 4.58579 35.6681C4.21071 35.293 4 34.7843 4 34.2539V20.2539H36V34.2539ZM36 16.2539H4V10.2539C4 9.72347 4.21071 9.21477 4.58579 8.83969C4.96086 8.46462 5.46957 8.25391 6 8.25391H10V10.2539C10 10.7843 10.2107 11.293 10.5858 11.6681C10.9609 12.0432 11.4696 12.2539 12 12.2539C12.5304 12.2539 13.0391 12.0432 13.4142 11.6681C13.7893 11.293 14 10.7843 14 10.2539V8.25391H26V10.2539C26 10.7843 26.2107 11.293 26.5858 11.6681C26.9609 12.0432 27.4696 12.2539 28 12.2539C28.5304 12.2539 29.0391 12.0432 29.4142 11.6681C29.7893 11.293 30 10.7843 30 10.2539V8.25391H34C34.5304 8.25391 35.0391 8.46462 35.4142 8.83969C35.7893 9.21477 36 9.72347 36 10.2539V16.2539Z"
-      fill={fill}
-    />
-  </svg>
-);
-
-const SidebarProfileIcon: SidebarIconComponent = ({ stroke }) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 32 39"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M29.4168 36.5039V32.6706C29.4168 30.6372 28.6091 28.6872 27.1713 27.2494C25.7335 25.8116 23.7835 25.0039 21.7502 25.0039H10.2502C8.21684 25.0039 6.26679 25.8116 4.82901 27.2494C3.39123 28.6872 2.5835 30.6372 2.5835 32.6706V36.5039"
-      stroke={stroke}
-      strokeWidth="3.83333"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-    <path
-      d="M16.0002 17.3372C20.2343 17.3372 23.6668 13.9048 23.6668 9.67057C23.6668 5.43639 20.2343 2.00391 16.0002 2.00391C11.766 2.00391 8.3335 5.43639 8.3335 9.67057C8.3335 13.9048 11.766 17.3372 16.0002 17.3372Z"
-      stroke={stroke}
-      strokeWidth="3.83333"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-);
-
-
-export default function JapaneseFoodApp() {
-  const [currentScreen, setCurrentScreen] = useState<Screen>("dashboard");
-  const [hoveredNav, setHoveredNav] = useState<SidebarNavKey | null>(null);
-  const [monthlyBudget, setMonthlyBudget] = useState(7500);
-  const [products, setProducts] = useState<Product[]>(sampleProducts);
-  const catalogScrollRef = useRef<HTMLDivElement | null>(null);
-  const subscriptionScrollRef = useRef<HTMLDivElement | null>(null);
-  const pendingProductScrollTop = useRef<number | null>(null);
-  const [profilePage, setProfilePage] = useState(1);
-  const [landingPage, setLandingPage] = useState(1);
-  const totalLandingPages = LANDING_CARD_CONTENT_PAGES.length;
-  const landingPageIndex = Math.min(Math.max(landingPage - 1, 0), totalLandingPages - 1);
-  const currentLandingCards = LANDING_CARD_CONTENT_PAGES[landingPageIndex] ?? [];
-  const totalProfilePages = 2;
-  const topButtonIconSize = 126;
-  const cartItems = useMemo(
-    () => products.filter((product) => product.quantity > 0),
-    [products]
-  );
-  const catalogQuantitySum = useMemo(
-    () => products.reduce((sum, product) => sum + product.quantity, 0),
-    [products]
-  );
-  const catalogPriceSum = useMemo(
-    () => products.reduce((sum, product) => sum + product.quantity * product.price, 0),
-    [products]
-  );
-
-  useEffect(() => {
-    if (currentScreen !== "profile" && profilePage !== 1) {
-      setProfilePage(1);
-    }
-  }, [currentScreen, profilePage]);
-
-  const updateProductQuantity = (id: number, change: number) => {
-    const activeScrollContainer = catalogScrollRef.current ?? subscriptionScrollRef.current;
-    if (activeScrollContainer) {
-      pendingProductScrollTop.current = activeScrollContainer.scrollTop;
-    }
-    setProducts((items) =>
-    items.map((item) =>
-    item.id === id ?
-    { ...item, quantity: Math.max(0, item.quantity + change) } :
-    item
-    )
-    );
-  };
-
-  useLayoutEffect(() => {
-    const activeScrollContainer = catalogScrollRef.current ?? subscriptionScrollRef.current;
-    if (pendingProductScrollTop.current !== null && activeScrollContainer) {
-      activeScrollContainer.scrollTop = pendingProductScrollTop.current;
-      pendingProductScrollTop.current = null;
-    }
-  }, [products]);
-
-  const Sidebar = () => {
-    const renderSidebarButton = (
-      key: SidebarNavKey,
-      label: string,
-      Icon: SidebarIconComponent,
-      dataOid: string,
-      onClick?: () => void
-    ) => {
-      const isSubscriptionKey = key === "subscription";
-      const isCatalogKey = key === "catalog";
-      const isActive = isSubscriptionKey ?
-        currentScreen === "subscription" :
-        isCatalogKey ?
-        currentScreen === "catalog" || currentScreen === "catalogLanding" :
-        currentScreen === key;
-      const isHovered = hoveredNav === key;
-      const showHighlight = isActive || isHovered;
-      const highlightOpacity = isActive ? 1 : isHovered ? 0.4 : 0;
-      const iconFill = isActive ? "#209fde" : "#ffffff";
-      const iconStroke = isActive ? "#209fde" : "#ffffff";
-
-      return (
-        <Button
-          key={key}
-          variant="ghost"
-          size="sm"
-          className="w-full flex justify-end px-0 py-0 bg-transparent hover:bg-transparent shadow-none focus-visible:ring-0 focus-visible:ring-offset-0"
-          style={{ height: "99.27px" }}
-          onMouseEnter={() => setHoveredNav(key)}
-          onMouseLeave={() => setHoveredNav(null)}
-          onClick={onClick}
-          data-oid={dataOid}>
-
-          <span
-            className="flex items-center justify-end"
-            style={{
-              width: "100%",
-              height: "99.27px",
-              flexShrink: 0
-            }}>
-
-            <span
-              className="flex items-center relative"
-              style={{
-                width: "192px",
-                height: "99.27px",
-                flexShrink: 0,
-                backgroundRepeat: "no-repeat",
-                backgroundSize: "100% 100%",
-                backgroundPosition: "right center",
-                justifyContent: "flex-start",
-                paddingLeft: "20px",
-                gap: "12px",
-                transform: "translateX(1px)"
-              }}>
-
-              <span
-              style={{
-                position: "absolute",
-                inset: 0,
-                backgroundImage: showHighlight ? SIDEBAR_ACTIVE_BG : "none",
-                backgroundRepeat: "no-repeat",
-                backgroundSize: "100% 100%",
-                backgroundPosition: "center",
-                opacity: highlightOpacity,
-                pointerEvents: "none",
-                zIndex: 0
-              }}>
-              </span>
-
-              <span style={{ ...SIDEBAR_ICON_STYLE, position: "relative", zIndex: 1 }}>
-                <Icon fill={iconFill} stroke={iconStroke} />
-              </span>
-              <span
-              className="text-xs font-medium"
-              style={{
-                ...SIDEBAR_LABEL_BASE_STYLE,
-                ...(isActive ? SIDEBAR_LABEL_SELECTED_STYLE : {}),
-                color: isActive ? "#101010" : "#ffffff",
-                position: "relative",
-                zIndex: 1
-              }}>
-                {label}
-              </span>
-            </span>
-          </span>
-        </Button>
-      );
-    };
-
-    return (
-      <div
-        className="w-[232px] bg-[#fda900] flex flex-col items-center py-6 space-y-8 fixed left-0 top-0 h-full z-10 rounded-tr-[20px] rounded-br-[20px]"
-        data-oid="ehf5pv5">
-
-        {/* Logo Section */}
-        <div className="flex flex-col items-center" data-oid="pm140hy">
-          <img
-          src="/images/logo_kaitasu.png"
-          alt="かいたす"
-          className="h-[72px] w-[72px]"
-          data-oid="kaitasu-logo" />
-        </div>
-
-        <nav className="w-full flex flex-col items-end space-y-2 pr-0" data-oid="bglg0.p">
-          {renderSidebarButton("dashboard", "ホーム", SidebarHomeIcon, "lqzeqpl", () => setCurrentScreen("dashboard"))}
-          {renderSidebarButton("catalog", "買い出し", SidebarCatalogIcon, "iu4097_", () => setCurrentScreen("catalogLanding"))}
-          {renderSidebarButton("cart", "買い物かご", SidebarCartIcon, "1yqj0f6", () => setCurrentScreen("cart"))}
-          {renderSidebarButton("history", "購入履歴", SidebarHistoryIcon, "rzbvl29", () => setCurrentScreen("history"))}
-          {renderSidebarButton(
-            "subscription",
-            "定期購入",
-            SidebarSubscriptionIcon,
-            "sub-btn-1",
-            () => setCurrentScreen("subscription")
-          )}
-          {renderSidebarButton("profile", "マイページ", SidebarProfileIcon, "jyqaqbq", () => setCurrentScreen("profile"))}
-        </nav>
-      </div>
-    );
-  };
-
-
-  const Dashboard = () =>
-  <div
-    className="flex-1 bg-white p-6 ml-[232px] min-h-screen"
-    data-oid="ysuosjy">
-
-      <div
-      className="mx-auto w-[903px] space-y-4"
-      data-oid=":xr07ip">
-
-        {/* Notification Card */}
-        <Card
-        className="w-[903px] p-4"
-        style={{
-          borderRadius: "14.469px",
-          border: "5px solid var(--primary, #FDA900)",
-          background: "var(--card, #FFF)"
-        }}
-        data-oid="2bp9lo4">
-
-          <h3
-          className="text-center font-medium mb-3 text-sm"
-          data-oid="aek175l">
-
-            お知らせ
-          </h3>
-          <div className="h-12 bg-[#adadad] rounded" data-oid="v_esgva"></div>
-        </Card>
-
-        {/* Budget Card */}
-        <Card
-        className="w-[903px] h-[230px] p-4 gap-[9px]"
-        style={{
-          borderRadius: "14.469px",
-          border: "5px solid var(--primary, #FDA900)",
-          background: "var(--card, #FFF)"
-        }}
-        data-oid="t2fyzfc">
-
-          <div
-          className="flex justify-between  items-smart mt-[7px] mb-2"
-          data-oid=":be0a2k">
-
-            <span className="text-sm font-medium text-[30px]" data-oid=":6v2jl-">
-              今月の予算 ▶
-            </span>
-            <div className="flex items-baseline gap-2 mt-[29px] mr-[81px] self-end" data-oid="qq2w-88">
-              <span className="font-bold text-lg text-[20px]">残り</span>
-              <span className="font-bold text-[36px] leading-none">
-                ¥{monthlyBudget.toLocaleString("ja-JP")}
-              </span>
-            </div>
-          </div>
-          <div className="flex w-[703px] justify-between mb-[7px] mt-[9px]"  data-oid="h3m4f7m">
-          <div className="text-xs text-[#adadad] text-[20px]" data-oid="woecoz3">
-            ¥12,500
-          </div>
-          <div className="flex" data-oid="h3m4f7m">
-            <div className="text-xs text-right text-[#209fde] text-[20px]" data-oid="6x91r4a">
-            使用率:
-            </div>
-            <div className="text-xs text-right text-[#101010] text-[20px]" data-oid="6x91r4a">
-            60%
-          </div>
-          </div>
-          </div>
-          <div
-          className="w-[744px] bg-[#adadad] rounded-full h-[34px] mb-1 mx-auto"
-          data-oid="pk6di4e">
-
-            <div
-            className="bg-[#209fde] h-[34px] rounded-full"
-            style={{ width: "60%" }}
-            data-oid="5_0shjj">
-          </div>
-          </div>
-        </Card>
-
-        <div className="mx-auto flex justify-center gap-[61px] w-[903px]" data-oid="ta8f1it">
-          <Button
-          variant="outline"
-          className="w-[180px] h-[180px] text-center border-[5px] border-[#fda900] bg-white hover:bg-gray-50 rounded-[20px] flex-col px-6 pb-6 pt-[17px] relative"
-          style={{
-            filter: "drop-shadow(4.5px 4.5px 0 #E4E2E2)"
-          }}
-          onClick={() => setCurrentScreen("catalog")}
-          data-oid="6g1e2.x">
-
-            <div className="flex h-[145px] flex-col items-center justify-between">
-              <span
-              className="text-[#101010] font-['BIZ_UDPGothic'] text-[20px] font-bold leading-normal"
-              data-oid="1417ieu">
-
-                買い出し
-              </span>
-              <div className="flex w-full justify-center">
-                <div style={{ width: "126px", height: "126px" }}>
-                  <SidebarCatalogIcon fill="#209fde" stroke="#209fde" />
-                </div>
-              </div>
-            </div>
-          </Button>
-          <Button
-          variant="outline"
-          className="w-[180px] h-[180px] text-center border-[5px] border-[#fda900] bg-white hover:bg-gray-50 rounded-[20px] flex-col px-6 pb-6 pt-[17px] relative"
-          style={{
-            filter: "drop-shadow(4.5px 4.5px 0 #E4E2E2)"
-          }}
-          onClick={() => setCurrentScreen("catalog")}
-          data-oid="eljule7">
-
-            <div className="flex h-[145px] flex-col items-center justify-between">
-              <span
-              className="text-[#101010] font-['BIZ_UDPGothic'] text-[20px] font-bold leading-normal"
-              data-oid="_8lynit">
-
-                定期購入の確認
-              </span>
-
-              <div className="flex w-full justify-center">
-                <div style={{ width: "110px", height: "110px", padding: "10px" }}>
-                  <SidebarSubscriptionIcon fill="#209fde" stroke="#209fde" />
-                </div>
-              </div>
-            </div>
-          </Button>
-          <Button
-          variant="outline"
-          className="w-[180px] h-[180px] text-center border-[5px] border-[#fda900] bg-white hover:bg-gray-50 rounded-[20px] flex-col px-6 pb-6 pt-[17px] relative"
-          style={{
-            filter: "drop-shadow(4.5px 4.5px 0 #E4E2E2)"
-          }}
-          onClick={() => setCurrentScreen("history")}
-          data-oid="vhz4ae4">
-
-            <div className="flex h-[145px] flex-col items-center justify-between">
-              <span
-              className="text-[#101010] font-['BIZ_UDPGothic'] text-[20px] font-bold leading-normal"
-              data-oid="pu-k0wb">
-
-                購入履歴
-              </span>
-
-              <div className="flex w-full justify-center">
-                <div style={{ width: "91px", height: "91px" }}>
-                  <SidebarHistoryIcon fill="#209fde" stroke="#209fde" />
-                </div>
-              </div>
-            </div>
-          </Button>
-          <Button
-          variant="outline"
-          className="w-[180px] h-[180px] text-center border-[5px] border-[#fda900] bg-white hover:bg-gray-50 rounded-[20px] flex-col px-6 pb-6 pt-[17px] relative"
-          style={{
-            filter: "drop-shadow(4.5px 4.5px 0 #E4E2E2)"
-          }}
-          onClick={() => setCurrentScreen("profile")}
-          data-oid="oodndhw">
-
-            <div className="flex h-[145px] flex-col items-center justify-between">
-              <span
-              className="text-[#101010] font-['BIZ_UDPGothic'] text-[20px] font-bold leading-normal"
-              data-oid="5s21wmk">
-
-                マイページ
-              </span>
-
-              <div className="flex w-full justify-center">
-                <div
-                style={{
-                  width: "65.386px",
-                  height: "84.068px",
-                  flexShrink: 0
-                }}>
-                  <SidebarProfileIcon fill="#209fde" stroke="#209fde" />
-                </div>
-              </div>
-            </div>
-          </Button>
-        </div>
-      </div>
-    </div>;
-
-
-  const Catalog = () =>
-  <div
-    className="flex-1 bg-white p-6 ml-[232px] relative min-h-screen"
-    data-oid="d95y1m5">
-
-      <div
-      className="absolute right-0 top-0 bottom-0 w-1 bg-[#fda900]"
-      data-oid="kh0_yce">
-    </div>
-
-      <div className="mx-auto w-[1000px]" data-oid="psedc55">
-        <div className="relative mb-6" data-oid="gfk0oa_">
-          <Search
-          className="absolute left-4 top-1/2 -translate-y-1/2 text-[#fda900]"
-          size={18}
-          data-oid="wum_nxs" />
-
-
-          <Input
-          placeholder="商品名で検索"
-          className="pl-12 h-12 border-2 border-[#fda900] text-sm rounded-lg bg-white shadow-sm focus:border-[#209fde] focus:ring-2 focus:ring-[#209fde]/20"
-          data-oid="fsk2zzr" />
-
-        </div>
-
-          <div className="flex gap-[25px] mb-6" data-oid="br-d9o3">
-            <Button
-              key="catalog-back-button"
-              variant="ghost"
-              className="border border-transparent p-0"
-              style={FILTER_BUTTON_INACTIVE_STYLE}
-              onClick={() => setCurrentScreen("catalogLanding")}
-              data-oid="catalog-back-button">
-              <span className="mr-3 flex items-center">
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  width="48"
-                  height="48"
-                  viewBox="0 0 48 48"
-                  fill="none">
-                  <path
-                    fillRule="evenodd"
-                    clipRule="evenodd"
-                    d="M22.4404 3.42066C21.5311 3.42066 20.659 3.78188 20.016 4.42486C19.373 5.06785 19.0118 5.93992 19.0118 6.84923C19.0118 7.75854 19.373 8.63062 20.016 9.2736C20.659 9.91658 21.5311 10.2778 22.4404 10.2778H31.7935C34.1613 10.3239 36.4166 11.2969 38.0749 12.9877C39.7331 14.6786 40.662 16.9524 40.662 19.3207C40.662 21.6889 39.7331 23.9627 38.0749 25.6536C36.4166 27.3444 34.1613 28.3174 31.7935 28.3635H16.5158V22.4389C16.5157 21.7609 16.3145 21.0982 15.9377 20.5345C15.561 19.9708 15.0256 19.5315 14.3992 19.272C13.7728 19.0126 13.0835 18.9447 12.4186 19.0769C11.7536 19.2092 11.1427 19.5356 10.6632 20.0149L1.31694 29.3612C0.835444 29.8396 0.506738 30.45 0.372461 31.1153C0.238184 31.7806 0.304378 32.4707 0.562658 33.0984C0.729515 33.5098 0.978658 33.8835 1.31009 34.2195L10.6632 43.5727C11.1427 44.052 11.7536 44.3784 12.4186 44.5107C13.0835 44.6429 13.7728 44.575 14.3992 44.3156C15.0256 44.0561 15.561 43.6168 15.9377 43.0531C16.3145 42.4894 16.5157 41.8267 16.5158 41.1487V35.2207H31.7935C33.9012 35.2516 35.994 34.8632 37.9502 34.0781C39.9064 33.2929 41.687 32.1266 43.1884 30.6471C44.6898 29.1676 45.882 27.4043 46.6958 25.4599C47.5096 23.5154 47.9287 21.4286 47.9287 19.3207C47.9287 17.2128 47.5096 15.1259 46.6958 13.1814C45.882 11.237 44.6898 9.47372 43.1884 7.9942C41.687 6.51468 39.9064 5.34843 37.9502 4.56327C35.994 3.77811 33.9012 3.38971 31.7935 3.42066H22.4404Z"
-                    fill="#FDA900"
-                  />
-                </svg>
-              </span>
-              <span style={FILTER_BUTTON_TEXT_STYLE}>戻る</span>
-            </Button>
-            {CATALOG_FILTER_BUTTONS.map(({ label, buttonDataOid, textDataOid }) => (
-              <Button
-                key={label}
-                variant="ghost"
-                className="border border-transparent p-0"
-                style={FILTER_BUTTON_INACTIVE_STYLE}
-                data-oid={buttonDataOid}>
-                <span style={FILTER_BUTTON_TEXT_STYLE} data-oid={textDataOid}>
-                  {label}
-              </span>
-            </Button>
-          ))}
-        </div>
-
-        <div
-        className="relative mt-[23px] mb-[24px] w-[1000px]"
-        data-oid="catalog-card-background">
-          <div
-          className="pointer-events-none absolute top-0 left-0 z-0 h-[507px] w-[1000px]"
-          style={{
-            borderRadius: "20px",
-            background: "rgba(253, 169, 0, 0.5)"
-          }}
-          aria-hidden="true"
-          />
-          <div
-          ref={catalogScrollRef}
-          className="relative z-10 h-[507px] w-[1000px] overflow-x-hidden overflow-y-auto pt-[18px] flex flex-col items-center"
-          data-oid="uy6_dcp-wrapper">
-            <div
-            className="flex justify-center gap-[25px] mb-6"
-            data-oid="uy6_dcp">
-            <span
-            className="flex items-center"
-            style={{
-              color: "var(--, #101010)",
-              fontFamily: '"BIZ UDPGothic"',
-              fontSize: "32px",
-              fontStyle: "normal",
-              fontWeight: 700,
-              lineHeight: "normal",
-              letterSpacing: "1.664px"
-            }}
-            data-oid="catalog-page-label">
-              ページ
-            </span>
-            {[1, 2, 3, 4, 5].map((num) => {
-              const isActive = num === 1;
-              return (
-                <Button
-                  key={num}
-                  size="sm"
-                  variant="ghost"
-                  className="border border-transparent p-0"
-                  style={{
-                    display: "flex",
-                    width: "60px",
-                    height: "60px",
-                    padding: "14px 19px",
-                    flexDirection: "column",
-                    justifyContent: "center",
-                    alignItems: "center",
-                    gap: "10px",
-                    flexShrink: 0,
-                    borderRadius: "20px",
-                    background: isActive ? "var(--, #FDA900)" : "var(--, #FFF)",
-                    backgroundColor: isActive ? "#FDA900" : "#FFF",
-                    boxShadow: isActive ?
-                      "0 4px 4px 0 rgba(0, 0, 0, 0.25) inset" :
-                      "0 4px 4px 0 rgba(0, 0, 0, 0.25)",
-                    color: "var(--, #101010)",
-                    fontFamily: '"BIZ UDPGothic"',
-                    fontSize: "32px",
-                    fontStyle: "normal",
-                    fontWeight: 700,
-                    lineHeight: "normal",
-                    letterSpacing: "1.664px"
-                  }}
-                  data-oid="9r7o0ii">
-                    {num}
-                  </Button>
-              );
-            })}
-            <span className="text-sm self-center font-medium" data-oid="xd09rri">
-              ...
-            </span>
-            <Button
-            size="sm"
-            variant="ghost"
-            className="border border-transparent p-0"
-            style={{
-              display: "flex",
-              width: "60px",
-              height: "60px",
-              padding: "14px 19px",
-              flexDirection: "column",
-              justifyContent: "center",
-              alignItems: "center",
-              gap: "10px",
-              borderRadius: "20px",
-              background: "var(--, #FFF)",
-              backgroundColor: "#FFF",
-              boxShadow: "0 4px 4px 0 rgba(0, 0, 0, 0.25)",
-              color: "var(--, #101010)",
-              fontFamily: '"BIZ UDPGothic"',
-              fontSize: "32px",
-              fontStyle: "normal",
-              fontWeight: 700,
-              lineHeight: "normal",
-              letterSpacing: "1.664px"
-            }}
-            data-oid="f8qlkjv">
-              15
-            </Button>
-            </div>
-
-            <div
-            className="grid grid-cols-4 gap-y-4"
-            style={{
-              columnGap: "45px",
-              gridTemplateColumns: "repeat(4, 188px)",
-              justifyItems: "start"
-            }}
-            data-oid="h7qwqv1">
-            {products.map((product) =>
-          <Card
-            key={product.id}
-            className="px-4 pt-1 pb-0 bg-white border-2 border-[#e0e0e0] rounded-xl shadow-sm hover:shadow-md transition-shadow flex flex-col"
-            style={{ width: "188px", height: "265px" }}
-            data-oid="3w-11ql">
-
-                <div className="flex justify-end mb-[3px]" data-oid="catalog-heart-row">
-                  <svg
-                xmlns="http://www.w3.org/2000/svg"
-                width="34"
-                height="31"
-                viewBox="0 0 34 31"
-                fill="none"
-                data-oid="eo3z5j5">
-                    <path
-                  d="M17 7.84292C17 7.84292 17 7.66699 15.7967 6.08366C14.4033 4.24699 12.345 2.91699 9.875 2.91699C5.9325 2.91699 2.75 6.09949 2.75 10.042C2.75 11.5145 3.19333 12.8762 3.95333 14.0003C5.23583 15.9162 17 28.2503 17 28.2503M17 7.84292C17 7.84292 17 7.66699 18.2033 6.08366C19.5967 4.24699 21.655 2.91699 24.125 2.91699C28.0675 2.91699 31.25 6.09949 31.25 10.042C31.25 11.5145 30.8067 12.8762 30.0467 14.0003C28.7642 15.9162 17 28.2503 17 28.2503"
-                  stroke="#209FDE"
-                  strokeWidth="3.86"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                />
-                  </svg>
-                </div>
-                <div
-                className="relative mb-[6.5px]"
-                style={{ height: "106px" }}
-                data-oid="8gipz64">
-                  <img
-                src={product.image || "/placeholder.svg"}
-                alt={product.name}
-                className="h-full w-full object-cover rounded-lg"
-                data-oid="q.-_y:_" />
-                </div>
-                <h4 className="text-sm font-bold" data-oid="d:mf1eo">
-                  {product.name}
-                </h4>
-                <p className="hidden" data-oid="ms.1o9h">
-                  {product.description}
-                </p>
-                <div
-              className="flex flex-col items-center"
-              data-oid="r7z2qp8">
-                <span
-                style={{
-                  color: "var(--, #101010)",
-                  fontFamily: '"BIZ UDPGothic"',
-                  fontSize: "20px",
-                  fontStyle: "normal",
-                  fontWeight: 700,
-                  lineHeight: "normal",
-                  letterSpacing: "1.04px",
-                  alignSelf: "flex-start",
-                  width: "100%",
-                  display: "block",
-                  marginTop: "12px"
-                }}
-                data-oid="eq4gt6a">
-                    ¥{product.price}
-                  </span>
-                  <div className="mt-[12px] flex w-full items-center justify-center gap-2" data-oid="4z4g4a-">
-                    <Button
-                  size="icon"
-                  variant="ghost"
-                  className="w-7 h-7 rounded-md bg-transparent hover:bg-transparent"
-                  onClick={() => updateProductQuantity(product.id, -1)}
-                  data-oid="les5kg0">
-                      <Image
-                    src="/images/mainasu.png"
-                    alt="数量を減らす"
-                    width={28}
-                    height={28}
-                    className="h-full w-full object-contain"
-                    data-oid="d7ycg-q" />
-
-                    </Button>
-                    <div
-                  className="flex items-center justify-center text-sm font-medium"
-                  style={{
-                    width: "73px",
-                    height: "26px",
-                    flexShrink: 0,
-                    borderRadius: "5px",
-                    border: "1px solid #FDA900",
-                    background: "#FFF"
-                  }}
-                  data-oid="f1loo1.">
-                      {product.quantity}
-                    </div>
-                    <Button
-                  size="icon"
-                  variant="ghost"
-                  className="w-7 h-7 rounded-md bg-transparent hover:bg-transparent"
-                  onClick={() => updateProductQuantity(product.id, 1)}
-                  data-oid=":--ycbg">
-                      <Image
-                    src="/images/plus.png"
-                    alt="数量を増やす"
-                    width={28}
-                    height={28}
-                    className="h-full w-full object-contain"
-                    data-oid="olbxiee" />
-
-                    </Button>
-                  </div>
-                </div>
-              </Card>
-          )}
-            </div>
-          </div>
-        </div>
-
-        {/* Cart Summary */}
-        <div
-        className="flex justify-between items-center bg-[#fda900] text-white p-4 rounded-xl border-2 border-[#fda900] shadow-md"
-        data-oid="t2ub1nx">
-
-          <span className="text-base font-bold" data-oid="5rigjnl">
-            {catalogQuantitySum}点 ¥
-            {catalogPriceSum}
-          </span>
-          <Button
-          className="bg-white text-[#fda900] text-base font-bold px-8 h-11 border-2 border-white rounded-lg hover:bg-gray-50 shadow-sm"
-          onClick={() => setCurrentScreen("cart")}
-          data-oid="b9c3xju">
-
-            購入
-          </Button>
-        </div>
-      </div>
-    </div>;
-
-
-  const Cart = () => {
-    const totalCartQuantity = cartItems.reduce(
-      (sum, item) => sum + item.quantity,
-      0
-    );
-    const totalCartPrice = cartItems.reduce(
-      (sum, item) => sum + item.price * item.quantity,
-      0
-    );
-
-    return (
-  <div className="flex-1 bg-white p-6 ml-[232px]" data-oid="qie1-gm">
-      <div className="max-w-sm mx-auto" data-oid="-j93.-d">
-        <h2 className="text-base font-bold mb-4" data-oid="gdwztq6">
-          買い物かご
-        </h2>
-
-        <div className="space-y-3 mb-6" data-oid=":2r1yrd">
-          {cartItems.map((item, index) =>
-        <Card
-          key={item.id}
-          className="p-3 bg-white border-2 border-gray-200 rounded-lg"
-          data-oid="5rgf5dd">
-
-              <div className="flex items-center gap-3" data-oid=".a7yvq5">
-                <img
-              src={item.image || "/placeholder.svg"}
-              alt={item.name}
-              className="w-12 h-12 object-cover rounded"
-              data-oid="b0g5711" />
-
-
-                <div className="flex-1" data-oid="yd66:be">
-                  <h4 className="text-sm font-medium" data-oid="5tlgvsv">
-                    {item.name}
-                  </h4>
-                  <p className="text-xs text-[#adadad]" data-oid="_b89cis">
-                    {item.description}
-                  </p>
-                </div>
-                <div className="text-right" data-oid="f:g9-vr">
-                  <div className="text-xs mb-1" data-oid="3_9wxay">
-                    数量
-                  </div>
-                  <div className="flex items-center gap-1" data-oid="zdp13dt">
-                  <Button
-                  size="icon"
-                  variant="ghost"
-                  className="w-5 h-5 rounded bg-transparent"
-                  onClick={() => updateProductQuantity(item.id, -1)}
-                  data-oid="g4qqy_k">
-                      <Image
-                    src="/images/mainasu.png"
-                    alt="数量を減らす"
-                    width={20}
-                    height={20}
-                    className="h-full w-full object-contain"
-                    data-oid="2h.nwkc" />
-                    </Button>
-                    <div
-                  className="flex items-center justify-center text-xs"
-                  style={{
-                    width: "73px",
-                    height: "26px",
-                    flexShrink: 0,
-                    borderRadius: "5px",
-                    border: "1px solid #FDA900",
-                    background: "#FFF"
-                  }}
-                  data-oid="weymacu">
-                      {item.quantity}
-                    </div>
-                  <Button
-                  size="icon"
-                  variant="ghost"
-                  className="w-5 h-5 rounded bg-transparent"
-                  onClick={() => updateProductQuantity(item.id, 1)}
-                  data-oid="4kc4n3t">
-                      <Image
-                    src="/images/plus.png"
-                    alt="数量を増やす"
-                    width={20}
-                    height={20}
-                    className="h-full w-full object-contain"
-                    data-oid=".gef-yq" />
-                    </Button>
-                  </div>
-                </div>
-                <div className="text-right" data-oid="8ke_pvk">
-                  <span className="font-bold text-sm" data-oid="jo1znli">
-                    ¥{item.price}
-                  </span>
-                </div>
-              </div>
-            </Card>
-        )}
-        </div>
-
-        <div
-        className="border-t-2 border-gray-200 pt-4 mb-6"
-        data-oid="lx.9fs:">
-
-          <div
-          className="flex justify-between text-base font-bold"
-          data-oid="e-466ve">
-
-            <span data-oid="qr0:wva">合計 {totalCartQuantity}点</span>
-            <span data-oid="5tmrc4s">
-              ¥
-              {totalCartPrice}
-            </span>
-          </div>
-        </div>
-
-        <div className="flex gap-3" data-oid="ufdhcq:">
-          <Button
-          variant="outline"
-          className="flex-1 text-sm border-2 border-[#fda900] text-[#fda900] rounded-md bg-transparent"
-          onClick={() => setCurrentScreen("order")}
-          data-oid="8i_h.o.">
-
-            注文確認
-          </Button>
-          <Button
-          className="flex-1 bg-[#fda900] text-sm border-2 border-[#fda900] rounded-md hover:bg-[#fda900]/90"
-          onClick={() => setCurrentScreen("catalog")}
-          data-oid="mgchczd">
-
-            買い物を続ける
-          </Button>
-        </div>
-      </div>
-    </div>
-    );
-  };
-
-
-  const Order = () =>
-  <div className="flex-1 bg-white p-6 ml-[232px]" data-oid="syv6qb8">
-      <div className="max-w-sm mx-auto" data-oid="9cyq3uk">
-        <h2 className="text-base font-bold mb-4" data-oid="ycsxz70">
-          注文確認
-        </h2>
-
-        {/* Delivery Address */}
-        <Card
-        className="p-4 mb-4 bg-white border-2 border-[#fda900] rounded-lg"
-        data-oid="jk:2_ry">
-
-          <div className="flex items-start justify-between" data-oid="ih9dsxw">
-            <div data-oid="r6l2mt0">
-              <h3
-              className="font-medium mb-2 text-sm flex items-center"
-              data-oid="fp4z1m-">
-
-                <span className="mr-2" data-oid="qyv7-tr">
-                  📍
-                </span>
-                配達先
-              </h3>
-              <p className="text-sm" data-oid="itribae">
-                徳島県鳴門市○○○
-              </p>
-              <p className="text-sm" data-oid=".5ob1_r">
-                西口○○
-              </p>
-            </div>
-            <Button
-            size="sm"
-            className="bg-[#fda900] text-white text-xs border-2 border-[#fda900] rounded-md hover:bg-[#fda900]/90"
-            data-oid="5zhtpvt">
-
-              変更
-            </Button>
-          </div>
-        </Card>
-
-        {/* Order Summary */}
-        <Card
-        className="p-4 mb-4 bg-white border-2 border-[#fda900] rounded-lg"
-        data-oid="8zvem9v">
-
-          <div className="space-y-2" data-oid=":x26-uf">
-            <div className="flex justify-between text-sm" data-oid="0qqfqwa">
-              <span data-oid="s1w10pe">小計</span>
-              <span data-oid="vizf-74">¥5,000</span>
-            </div>
-            <div className="flex justify-between text-sm" data-oid="u9:dvh7">
-              <span data-oid="-7g8x72">配送料</span>
-              <span data-oid="z4m-3lx">¥100</span>
-            </div>
-            <div
-            className="border-t-2 border-gray-200 pt-2 flex justify-between font-bold text-sm"
-            data-oid=".dfs9ab">
-
-              <span data-oid="7b245_-">合計</span>
-              <span data-oid="21xx5:h">¥5,100</span>
-            </div>
-          </div>
-        </Card>
-
-        {/* Delivery Schedule */}
-        <Card
-        className="p-4 mb-6 bg-white border-2 border-[#fda900] rounded-lg"
-        data-oid="a13j9dc">
-
-          <h3
-          className="font-medium mb-2 text-sm flex items-center"
-          data-oid="qfwuago">
-
-            <span className="mr-2" data-oid="riz.ufg">
-              📅
-            </span>
-            配達メモ
-          </h3>
-          <p className="text-sm" data-oid="zeyhnax">
-            2025年 9月1日午後16:00
-          </p>
-          <p className="text-sm" data-oid=":zrymkx">
-            ご注文予定
-          </p>
-        </Card>
-
-        {/* Action Buttons */}
-        <div className="flex gap-3" data-oid="nz7v1jm">
-          <Button
-          variant="outline"
-          className="flex-1 text-sm border-2 border-[#fda900] text-[#fda900] rounded-md bg-transparent"
-          onClick={() => setCurrentScreen("cart")}
-          data-oid="-fxs0ta">
-
-            かごに戻る
-          </Button>
-          <Button
-          className="flex-1 bg-[#fda900] text-white text-sm border-2 border-[#fda900] rounded-md hover:bg-[#fda900]/90"
-          data-oid="khc3gs1">
-
-            注文を確定してサイトへ
-          </Button>
-        </div>
-      </div>
-    </div>;
-
-
-  const History = () =>
-  <div className="flex-1 bg-white p-6 ml-[232px]" data-oid="n3s1kmy">
-      <div className="max-w-sm mx-auto" data-oid="me-dwvn">
-        <h2 className="text-base font-bold mb-4" data-oid="x::kret">
-          購入履歴
-        </h2>
-
-        <div className="space-y-3" data-oid="4c1ea9l">
-          {Array.from({ length: 6 }).map((_, i) =>
-        <Card
-          key={i}
-          className="p-3 bg-white border-2 border-gray-200 rounded-lg"
-          data-oid="giqzpfm">
-
-              <div className="flex items-center gap-3" data-oid="jlp_i_y">
-                <div className="text-xs" data-oid="jflct6b">
-                  <div data-oid="mtapic7">2025年</div>
-                  <div data-oid="op7puzz">9月1日</div>
-                </div>
-                <img
-              src="/images/food-item.jpg"
-              alt="トップリブ 特製"
-              className="w-12 h-12 object-cover rounded"
-              data-oid="wur3l8e" />
-
-
-                <div className="flex-1" data-oid="kmj3p4f">
-                  <h4 className="text-sm font-medium" data-oid="uwyj:6y">
-                    トップバリュ 特製
-                  </h4>
-                  <p className="text-xs text-[#adadad]" data-oid="p5nlqyv">
-                    ナンバー 260g
-                  </p>
-                  <Badge
-                variant="secondary"
-                className="text-xs border border-gray-300 rounded"
-                data-oid="_ne7wd0">
-
-                    お気に入りに追加
-                  </Badge>
-                </div>
-                <div className="text-right" data-oid="t6t6kx-">
-                  <div className="text-xs" data-oid="9d22m2g">
-                    1コ
-                  </div>
-                  <div className="font-bold text-sm" data-oid="8yen38r">
-                    ¥350
-                  </div>
-                </div>
-              </div>
-            </Card>
-        )}
-        </div>
-      </div>
-    </div>;
-
-
-  const Profile = () =>
-  <div className="flex-1 bg-white p-6 ml-[232px]" data-oid="3pchgx4">
-      <div className="flex flex-col items-center gap-4" data-oid=":530dgu">
-
-        <Card
-        className="flex flex-col gap-6 p-6"
-        style={{
-          width: "900px",
-          height: "633px",
-          flexShrink: 0,
-          borderRadius: "14.469px",
-          border: "5px solid #FDA900",
-          backgroundColor: "#FFF"
-        }}
-        data-oid="profile-main-card">
-
-          <div className="flex items-center gap-3" data-oid="mhd5qso">
-            {profilePage === 1 ? (
-              <>
-                <div
-                className="w-12 h-12 bg-[#adadad] rounded-full"
-                data-oid="v96ohmr">
-              </div>
-              <div data-oid="w32:5lp" className="flex items-center gap-4">
-                <div className="flex items-baseline gap-1" data-oid="2y3hjo1">
-                  <span
-                    style={{
-                      color: "var(--, #101010)",
-                      fontFamily: "BIZ UDPGothic",
-                      fontSize: "36px",
-                      fontStyle: "normal",
-                      fontWeight: 700,
-                      lineHeight: "normal"
-                    }}
-                  >
-                    水口 和佳
-                  </span>
-                  <span
-                    style={{
-                      color: "var(--, #101010)",
-                      fontFamily: "BIZ UDPGothic",
-                      fontSize: "24px",
-                      fontStyle: "normal",
-                      fontWeight: 700,
-                      lineHeight: "normal"
-                    }}
-                  >
-                    さん
-                  </span>
-                </div>
-                <Button
-                variant="ghost"
-                className="border border-transparent p-0"
-                style={FILTER_BUTTON_INACTIVE_STYLE}
-                data-oid="541gvwr">
-
-                  <span style={FILTER_BUTTON_TEXT_STYLE}>
-                    名前を変更
-                  </span>
-                </Button>
-              </div>
-            </>
-            ) : null}
-          </div>
-
-          <div className="flex-1 w-full" data-oid="o4a2j0u">
-            {profilePage === 1 ? (
-              <div className="space-y-6" data-oid="profile-page-one">
-                <div>
-                  <h3 className="font-medium mb-3 text-sm" data-oid="xkxxiap">
-                    予算設定
-                  </h3>
-                  <div className="space-y-3" data-oid="prr6v6o">
-                    <div data-oid="zve5:m0">
-                      <label className="text-sm block mb-1" data-oid="d68qzu_">
-                        今月の予算
-                      </label>
-                      <div className="flex items-center gap-2" data-oid="lirbdor">
-                        <span className="text-sm" data-oid="zr1baos">
-                          ¥
-                        </span>
-                        <Input
-                          type="number"
-                          value={monthlyBudget}
-                          onChange={(e) => setMonthlyBudget(Number(e.target.value))}
-                          className="flex-1 text-sm border-2 border-gray-300 rounded-md"
-                          data-oid="_x89egx"
-                        />
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div>
-                  <h3 className="font-medium mb-3 text-sm" data-oid="calc-period-label">
-                    計算期間
-                  </h3>
-                  <div className="flex flex-wrap" style={{ gap: "37px" }} data-oid="calc-period-buttons">
-                    {[1, 2, 3, 4, 5, 6].map((num) => (
-                      <Button
-                        key={`calc-period-${num}`}
-                        variant="ghost"
-                        className="border border-transparent p-0"
-                        style={FILTER_BUTTON_INACTIVE_STYLE}
-                        data-oid={`calc-period-${num}`}>
-                        <span style={FILTER_BUTTON_TEXT_STYLE}>期間 {num}</span>
-                      </Button>
-                    ))}
-                  </div>
-                </div>
-              </div>
-            ) : (
-              <div
-                className="flex h-full items-start"
-                data-oid="profile-page-two">
-                <div className="flex flex-col items-start gap-8 w-full max-w-[520px]" data-oid="profile-page-two-content">
-                  <div className="flex flex-col items-start gap-6" data-oid="profile-page-two-subsection">
-                    <p
-                      style={{
-                        color: "var(--, #101010)",
-                        fontFamily: '"BIZ UDPGothic"',
-                        fontSize: "24px",
-                        fontStyle: "normal",
-                        fontWeight: 700,
-                        lineHeight: "normal",
-                        letterSpacing: "1.248px"
-                      }}
-                      data-oid="profile-page-two-title">
-                      定期購入の確認
-                    </p>
-                    <Button
-                      variant="ghost"
-                      className="border border-transparent p-0"
-                      style={FILTER_BUTTON_INACTIVE_STYLE}
-                      data-oid="profile-page-two-button">
-                      <span style={FILTER_BUTTON_TEXT_STYLE}>定期購入の確認</span>
-                    </Button>
-                  </div>
-                  <div className="flex flex-col items-start gap-6" data-oid="profile-page-two-favorite-subsection">
-                    <p
-                      style={{
-                        color: "var(--, #101010)",
-                        fontFamily: '"BIZ UDPGothic"',
-                        fontSize: "24px",
-                        fontStyle: "normal",
-                        fontWeight: 700,
-                        lineHeight: "normal",
-                        letterSpacing: "1.248px"
-                      }}
-                      data-oid="profile-page-two-favorite-title">
-                      お気に入り登録の確認
-                    </p>
-                    <Button
-                      variant="ghost"
-                      className="border border-transparent p-0"
-                      style={FILTER_BUTTON_INACTIVE_STYLE}
-                      data-oid="profile-page-two-favorite-button">
-                      <span style={FILTER_BUTTON_TEXT_STYLE}>お気に入り一覧へ</span>
-                    </Button>
-                  </div>
-                </div>
-              </div>
-            )}
-          </div>
-
-          <div
-          className="mt-auto flex flex-col items-center gap-[24px]"
-          data-oid="profile-pagination-wrapper">
-            <div
-            className="flex items-center justify-center gap-[25px]"
-            data-oid="profile-pagination">
-              <span
-              style={{
-                color: "var(--, #101010)",
-                fontFamily: '"BIZ UDPGothic"',
-                fontSize: "32px",
-                fontStyle: "normal",
-                fontWeight: 700,
-                lineHeight: "normal",
-                letterSpacing: "1.664px"
-              }}
-              data-oid="profile-page-label">
-                ページ
-              </span>
-              {Array.from({ length: totalProfilePages }, (_, i) => i + 1).map((num) => {
-                const isActive = num === profilePage;
-                return (
-                  <Button
-                    key={`profile-page-${num}`}
-                    size="sm"
-                    variant="ghost"
-                    className="border border-transparent p-0"
-                    style={{
-                      display: "flex",
-                      width: "60px",
-                      height: "60px",
-                      padding: "14px 19px",
-                      flexDirection: "column",
-                      justifyContent: "center",
-                      alignItems: "center",
-                      gap: "10px",
-                      flexShrink: 0,
-                      borderRadius: "20px",
-                      background: isActive ? "var(--, #FDA900)" : "var(--, #FFF)",
-                      backgroundColor: isActive ? "#FDA900" : "#FFF",
-                      boxShadow: isActive ?
-                        "0 4px 4px 0 rgba(0, 0, 0, 0.25) inset" :
-                        "0 4px 4px 0 rgba(0, 0, 0, 0.25)",
-                      color: "var(--, #101010)",
-                      fontFamily: '"BIZ UDPGothic"',
-                      fontSize: "32px",
-                      fontStyle: "normal",
-                      fontWeight: 700,
-                      lineHeight: "normal",
-                      letterSpacing: "1.664px"
-                    }}
-                    onClick={() => setProfilePage(num)}
-                    data-oid={`profile-page-${num}`}>
-                    {num}
-                  </Button>
-                );
-              })}
-            </div>
-          </div>
-        </Card>
-      </div>
-    </div>;
-
-
-  const CatalogLanding = () =>
-  <div
-    className="flex-1 bg-white p-6 ml-[232px] min-h-screen"
-    data-oid="catalog-landing">
-
-      <div className="mx-auto w-[1000px]">
-        <div className="relative mb-6" data-oid="catalog-landing-search">
-          <Search
-          className="absolute left-4 top-1/2 -translate-y-1/2 text-[#fda900]"
-          size={18}
-          data-oid="catalog-landing-search-icon" />
-
-
-          <Input
-          placeholder="商品名で検索"
-          className="pl-12 h-12 border-2 border-[#fda900] text-sm rounded-lg bg-white shadow-sm focus:border-[#209fde] focus:ring-2 focus:ring-[#209fde]/20"
-          data-oid="catalog-landing-search-input" />
-
-        </div>
-
-        <div className="flex gap-[25px] mb-6" data-oid="catalog-landing-filters">
-          {LANDING_FILTER_BUTTONS.map(({ label, buttonDataOid, textDataOid }) => (
-            <Button
-              key={`${label}-landing`}
-              variant="ghost"
-              className="border border-transparent p-0"
-              style={FILTER_BUTTON_INACTIVE_STYLE}
-              data-oid={`${buttonDataOid}-landing`}>
-              <span style={FILTER_BUTTON_TEXT_STYLE} data-oid={`${textDataOid}-landing`}>
-                {label}
-              </span>
-            </Button>
-          ))}
-        </div>
-
-        <div
-        className="relative mt-[23px] mb-[24px] w-[1000px]"
-        data-oid="catalog-landing-card-background">
-          <div
-          className="pointer-events-none absolute top-0 left-0 z-0 h-[507px] w-[1000px]"
-          style={{
-            borderRadius: "20px",
-            background: "rgba(253, 169, 0, 0.5)"
-          }}
-          aria-hidden="true"
-          />
-          <div
-          className="relative z-10 h-[507px] w-[1000px] overflow-hidden pt-[18px] flex flex-col items-center gap-[36px]"
-          data-oid="catalog-landing-pagination-wrapper">
-            <div
-            className="flex items-center justify-center gap-[25px]"
-            data-oid="catalog-landing-pagination">
-              <span
-              style={{
-                color: "var(--, #101010)",
-                fontFamily: '"BIZ UDPGothic"',
-                fontSize: "32px",
-                fontStyle: "normal",
-                fontWeight: 700,
-                lineHeight: "normal",
-                letterSpacing: "1.664px"
-              }}
-              data-oid="catalog-landing-page-label">
-                ページ
-              </span>
-              {Array.from({ length: totalLandingPages }, (_, i) => i + 1).map((num) => {
-                const isActive = num === landingPage;
-                return (
-                  <Button
-                    key={`landing-page-${num}`}
-                    size="sm"
-                    variant="ghost"
-                    className="border border-transparent p-0"
-                    style={{
-                      display: "flex",
-                      width: "60px",
-                      height: "60px",
-                      padding: "14px 19px",
-                      flexDirection: "column",
-                      justifyContent: "center",
-                      alignItems: "center",
-                      gap: "10px",
-                      flexShrink: 0,
-                      borderRadius: "20px",
-                      background: isActive ? "var(--, #FDA900)" : "var(--, #FFF)",
-                      backgroundColor: isActive ? "#FDA900" : "#FFF",
-                      boxShadow: isActive ?
-                        "0 4px 4px 0 rgba(0, 0, 0, 0.25) inset" :
-                        "0 4px 4px 0 rgba(0, 0, 0, 0.25)",
-                      color: "var(--, #101010)",
-                      fontFamily: '"BIZ UDPGothic"',
-                      fontSize: "32px",
-                      fontStyle: "normal",
-                      fontWeight: 700,
-                      lineHeight: "normal",
-                      letterSpacing: "1.664px"
-                    }}
-                    onClick={() => setLandingPage(num)}
-                    data-oid={`catalog-landing-page-${num}`}>
-                    {num}
-                  </Button>
-                );
-              })}
-            </div>
-            <div
-            className="grid grid-cols-4"
-            style={{ gap: "45px" }}
-            data-oid="catalog-landing-card-grid">
-              {currentLandingCards.map((card, index) => {
-                const Icon = card.renderIcon;
-                return (
-                  <Button
-                    key={`landing-card-${landingPage}-${index}`}
-                    variant="ghost"
-                    className="relative flex h-[160px] w-[160px] flex-col items-center pt-[13px] border border-transparent p-0 transition-colors duration-200 hover:bg-[#fda900]/20"
-                    onClick={() => setCurrentScreen("catalog")}
-                    data-oid={`catalog-landing-card-${index}`}>
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="165"
-                      height="165"
-                      viewBox="0 0 165 165"
-                      fill="none"
-                      className="absolute inset-0"
-                      data-oid={`catalog-landing-card-svg-${index}`}>
-                      <g filter="url(#catalog-card-filter)">
-                        <rect width="160" height="160" rx="13.3333" fill="#FDA900" />
-                        <rect x="4.44434" y="4.44434" width="151.111" height="151.111" rx="10" fill="white" />
-                      </g>
-                      <defs>
-                        <filter id="catalog-card-filter" x="0" y="0" width="164.5" height="164.5" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
-                          <feFlood floodOpacity="0" result="BackgroundImageFix" />
-                          <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha" />
-                          <feOffset dx="4.5" dy="4.5" />
-                          <feComposite in2="hardAlpha" operator="out" />
-                          <feColorMatrix type="matrix" values="0 0 0 0 0.895207 0 0 0 0 0.887003 0 0 0 0 0.887003 0 0 0 1 0" />
-                          <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_3977_578" />
-                          <feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow_3977_578" result="shape" />
-                        </filter>
-                      </defs>
-                    </svg>
-                    <div
-                      className="relative z-10 flex h-full w-full flex-col items-center pt-[13px]"
-                      data-oid={`catalog-landing-card-content-${index}`}>
-                      <span
-                      style={{
-                        color: "var(--, #101010)",
-                        fontFamily: '"BIZ UDPGothic"',
-                        fontSize: "24px",
-                        fontStyle: "normal",
-                        fontWeight: 700,
-                        lineHeight: "24px",
-                        letterSpacing: "1.248px"
-                      }}
-                      data-oid={`catalog-landing-card-title-${index}`}>
-                        {card.title}
-                      </span>
-                      <div className="flex w-full flex-1 items-center justify-center pt-[13px]">
-                        {Icon ? (
-                          <div
-                          className="flex h-[100px] w-[100px] items-center justify-center"
-                          data-oid={`catalog-landing-card-icon-${index}`}>
-                            <Icon />
-                          </div>
-                        ) : (
-                          <div
-                          className="h-[100px] w-[100px] rounded-lg border border-dashed border-[#cfcfcf] bg-[#f5f5f5]"
-                          data-oid={`catalog-landing-card-icon-${index}`}
-                          />
-                        )}
-                      </div>
-                    </div>
-                  </Button>
-                );
-              })}
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>;
-
-
-  const Subscription = () =>
-  <div
-    className="flex-1 bg-white p-6 ml-[232px] relative min-h-screen"
-    data-oid="subscription-page">
-
-      <div
-      className="absolute right-0 top-0 bottom-0 w-1 bg-[#fda900]"
-      data-oid="subscription-accent">
-    </div>
-
-      <div className="mx-auto w-[1000px]" data-oid="subscription-content">
-        <div className="relative mb-6" data-oid="subscription-search">
-          <Search
-          className="absolute left-4 top-1/2 -translate-y-1/2 text-[#fda900]"
-          size={18}
-          data-oid="subscription-search-icon" />
-
-
-          <Input
-          placeholder="商品名で検索"
-          className="pl-12 h-12 border-2 border-[#fda900] text-sm rounded-lg bg-white shadow-sm focus:border-[#209fde] focus:ring-2 focus:ring-[#209fde]/20"
-          data-oid="subscription-search-input" />
-
-        </div>
-
-        <div
-        className="relative mt-[23px] mb-[24px] w-[1000px]"
-        data-oid="subscription-card-background">
-          <div
-          className="pointer-events-none absolute top-0 left-0 z-0 h-[662px] w-[1000px]"
-          style={{
-            borderRadius: "20px",
-            background: "rgba(253, 169, 0, 0.5)"
-          }}
-          aria-hidden="true"
-          />
-          <div
-          ref={subscriptionScrollRef}
-          className="relative z-10 h-[662px] w-[1000px] overflow-x-hidden overflow-y-auto pt-[18px] flex flex-col items-center"
-          data-oid="subscription-card-wrapper">
-            <div
-            className="flex justify-center gap-[25px] mb-6"
-            data-oid="subscription-pagination">
-            {[1, 2, 3, 4, 5].map((num) => {
-              const isActive = num === 1;
-              return (
-                <Button
-                  key={`subscription-page-${num}`}
-                  size="sm"
-                  variant="ghost"
-                  className="border border-transparent p-0"
-                  style={{
-                    display: "flex",
-                    width: "60px",
-                    height: "60px",
-                    padding: "14px 19px",
-                    flexDirection: "column",
-                    justifyContent: "center",
-                    alignItems: "center",
-                    gap: "10px",
-                    flexShrink: 0,
-                    borderRadius: "20px",
-                    background: isActive ? "var(--, #FDA900)" : "var(--, #FFF)",
-                    backgroundColor: isActive ? "#FDA900" : "#FFF",
-                    boxShadow: isActive ?
-                      "0 4px 4px 0 rgba(0, 0, 0, 0.25) inset" :
-                      "0 4px 4px 0 rgba(0, 0, 0, 0.25)",
-                    color: "var(--, #101010)",
-                    fontFamily: '"BIZ UDPGothic"',
-                    fontSize: "32px",
-                    fontStyle: "normal",
-                    fontWeight: 700,
-                    lineHeight: "normal",
-                    letterSpacing: "1.664px"
-                  }}
-                  data-oid={`subscription-page-button-${num}`}>
-                    {num}
-                  </Button>
-              );
-            })}
-            <span className="text-sm self-center font-medium" data-oid="subscription-pagination-ellipsis">
-              ...
-            </span>
-            <Button
-            size="sm"
-            variant="ghost"
-            className="border border-transparent p-0"
-            style={{
-              display: "flex",
-              width: "60px",
-              height: "60px",
-              padding: "14px 19px",
-              flexDirection: "column",
-              justifyContent: "center",
-              alignItems: "center",
-              gap: "10px",
-              borderRadius: "20px",
-              background: "var(--, #FFF)",
-              backgroundColor: "#FFF",
-              boxShadow: "0 4px 4px 0 rgba(0, 0, 0, 0.25)",
-              color: "var(--, #101010)",
-              fontFamily: '"BIZ UDPGothic"',
-              fontSize: "32px",
-              fontStyle: "normal",
-              fontWeight: 700,
-              lineHeight: "normal",
-              letterSpacing: "1.664px"
-            }}
-            data-oid="subscription-pagination-last">
-              15
-            </Button>
-            </div>
-            <div
-            className="grid grid-cols-4 gap-y-4"
-            style={{
-              columnGap: "45px",
-              gridTemplateColumns: "repeat(4, 188px)",
-              justifyItems: "start"
-            }}
-            data-oid="subscription-card-grid">
-            {products.map((product) =>
-          <Card
-            key={`subscription-${product.id}`}
-            className="px-4 pt-1 pb-0 bg-white border-2 border-[#e0e0e0] rounded-xl shadow-sm hover:shadow-md transition-shadow flex flex-col"
-            style={{ width: "188px", height: "265px" }}
-            data-oid="subscription-card">
-
-                <div className="flex justify-end mb-2" data-oid="subscription-heart-row">
-                  <svg
-                xmlns="http://www.w3.org/2000/svg"
-                width="34"
-                height="31"
-                viewBox="0 0 34 31"
-                fill="none"
-                data-oid="subscription-heart">
-                    <path
-                  d="M17 7.84292C17 7.84292 17 7.66699 15.7967 6.08366C14.4033 4.24699 12.345 2.91699 9.875 2.91699C5.9325 2.91699 2.75 6.09949 2.75 10.042C2.75 11.5145 3.19333 12.8762 3.95333 14.0003C5.23583 15.9162 17 28.2503 17 28.2503M17 7.84292C17 7.84292 17 7.66699 18.2033 6.08366C19.5967 4.24699 21.655 2.91699 24.125 2.91699C28.0675 2.91699 31.25 6.09949 31.25 10.042C31.25 11.5145 30.8067 12.8762 30.0467 14.0003C28.7642 15.9162 17 28.2503 17 28.2503"
-                  stroke="#209FDE"
-                  strokeWidth="3.86"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                />
-                  </svg>
-                </div>
-                <div
-                className="relative mb-[6.5px]"
-                style={{ height: "106px" }}
-                data-oid="subscription-image">
-                  <img
-                src={product.image || "/placeholder.svg"}
-                alt={product.name}
-                className="h-full w-full object-cover rounded-lg"
-                data-oid="subscription-image-img" />
-                </div>
-                <h4 className="text-sm font-bold" data-oid="subscription-name">
-                  {product.name}
-                </h4>
-                <p className="hidden" data-oid="subscription-description">
-                  {product.description}
-                </p>
-                <div
-              className="mt-auto flex flex-col gap-0"
-              data-oid="subscription-price-wrapper">
-                  <div
-                className="w-full"
-                style={{
-                  color: "var(--, #101010)",
-                  fontFamily: '"BIZ UDPGothic"',
-                  fontSize: "20px",
-                  fontStyle: "normal",
-                  fontWeight: 700,
-                  lineHeight: "normal",
-                  letterSpacing: "1.04px",
-                  marginTop: "12px"
-                }}
-                data-oid="subscription-price">
-                    ¥{product.price}
-                  </div>
-                  <div
-                className="flex w-full justify-center"
-                data-oid="subscription-favorite-wrapper">
-                    <Button
-                  variant="ghost"
-                  className="p-0 bg-transparent hover:bg-transparent"
-                  data-oid="subscription-favorite-button">
-                      <Image
-                    src="/images/favorite.png"
-                    alt="お気に入りに追加"
-                    width={152}
-                    height={36}
-                    className="h-[36px] w-[152px] object-contain"
-                    data-oid="subscription-favorite-img" />
-                    </Button>
-                  </div>
-                </div>
-              </Card>
-          )}
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>;
-
-
-  const renderScreen = () => {
-    switch (currentScreen) {
-      case "dashboard":
-        return <Dashboard data-oid="ylfph:g" />;
-      case "catalog":
-        return <Catalog data-oid="wcl4:a0" />;
-      case "catalogLanding":
-        return <CatalogLanding data-oid="catalog-landing-screen" />;
-      case "cart":
-        return <Cart data-oid="13gj:ol" />;
-      case "order":
-        return <Order data-oid="h1.bjtn" />;
-      case "history":
-        return <History data-oid="3.1vred" />;
-      case "profile":
-        return <Profile data-oid="faqj1gy" />;
-      case "subscription":
-        return <Subscription data-oid="subscription-screen" />;
-      default:
-        return <Dashboard data-oid="fntgw1v" />;
-    }
-  };
-
-  return (
-    <div className="h-screen bg-white" data-oid="i339shy">
-      <Sidebar data-oid="7_f:5ci" />
-      {renderScreen()}
-    </div>);
-
+export default function Page() {
+  return <JapaneseFoodApp />;
 }
