+++ lib/data/sampleProducts.ts
@@ -1,0 +1,68 @@
+import type { Product } from "@/types/page";
+
+export const SAMPLE_PRODUCTS: Product[] = [
+  {
+    id: 1,
+    name: "トップリブ 特製",
+    description: "ナンバー 260g",
+    price: 350,
+    image: "/images/food-item.jpg",
+    quantity: 0
+  },
+  {
+    id: 2,
+    name: "ビーフステーキ",
+    description: "プレミアム 300g",
+    price: 450,
+    image: "/images/food-item.jpg",
+    quantity: 0
+  },
+  {
+    id: 3,
+    name: "チキンカツ",
+    description: "ジューシー 200g",
+    price: 280,
+    image: "/images/food-item.jpg",
+    quantity: 0
+  },
+  {
+    id: 4,
+    name: "ポークソテー",
+    description: "厚切り 250g",
+    price: 380,
+    image: "/images/food-item.jpg",
+    quantity: 0
+  },
+  {
+    id: 5,
+    name: "ハンバーグ",
+    description: "手作り 180g",
+    price: 320,
+    image: "/images/food-item.jpg",
+    quantity: 0
+  },
+  {
+    id: 6,
+    name: "サーモン焼き",
+    description: "ノルウェー産 220g",
+    price: 420,
+    image: "/images/food-item.jpg",
+    quantity: 0
+  },
+  {
+    id: 7,
+    name: "エビフライ",
+    description: "大ぶり 3本",
+    price: 390,
+    image: "/images/food-item.jpg",
+    quantity: 0
+  },
+  {
+    id: 8,
+    name: "唐揚げ",
+    description: "ジューシー 6個",
+    price: 300,
+    image: "/images/food-item.jpg",
+    quantity: 0
+  }
+];
